% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axislabel.data.frame}
\alias{axislabel.data.frame}
\title{Axis Label for Data Frame}
\usage{
\method{axislabel}{data.frame}(x, var, log = FALSE, ...)
}
\arguments{
\item{x}{data.frame}

\item{var}{item of interest}

\item{log}{whether this is for a log scale}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Axis label for data.frame.
}
\seealso{
Other axislabel: \code{\link{axislabel}}
}
\keyword{internal}
