% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axislabel.folded}
\alias{axislabel.folded}
\title{Axis Label for Folded}
\usage{
\method{axislabel}{folded}(x, var, log = FALSE, ...)
}
\arguments{
\item{x}{folded}

\item{var}{item of interest}

\item{log}{whether this is for a log scale}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Axis label for folded.
}
\keyword{internal}
