% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axislabel}
\alias{axislabel}
\title{Generic Axis Label}
\usage{
axislabel(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Generic axis label, with method for 'data.frame'.
}
\seealso{
Other generic functions: \code{\link{corsplom}},
  \code{\link{dens}}, \code{\link{metaplot}},
  \code{\link{scatter}}

Other axislabel: \code{\link{axislabel.data.frame}}
}
\keyword{internal}
