% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bargraph.R
\name{Bargraph}
\alias{Bargraph}
\title{Internal Function}
\usage{
Bargraph(source_tbl, study, group_var = NULL, type = "Sensitivity",
  median = "sens", lci = "sens.ci.lower", uci = "sens.ci.upper",
  weightmark = NULL)
}
\arguments{
\item{source_tbl}{Dataframe containing the meta-analysis data}

\item{study}{Index column for unique study identifier e.g. "Obelix 50 B.c."}

\item{type}{Type of Data shown (e.g. Sensitivity). Is X axis Lable}

\item{median}{Median}

\item{lci}{Lower boundary of confidence interval}

\item{uci}{Upper boundary of confidence interval}

\item{weightmark}{column with values for weightings}
}
\value{
none
}
\description{
Creates the actual forest plot. Called by Forest_Plot
}
