% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{boxplot.folded}
\alias{boxplot.folded}
\title{Boxplot Method for Folded}
\usage{
\method{boxplot}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{passed to \code{\link{boxplot.data.frame}}}
}
\description{
Boxplot for folded. Converts to data.frame with defined column attributes and calls data.frame method.
}
\examples{
library(fold)
data(eventsf)
boxplot(eventsf, SEX, WT, ref = 68)
}
\seealso{
Other bivariate plots: \code{\link{metaplot}},
  \code{\link{scatter_data_frame}}, \code{\link{scatter}}

Other boxplot: \code{\link{boxplot.data.frame}}
}
