% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{boxplot}
\alias{boxplot}
\alias{boxplot.data.frame}
\alias{boxplot.folded}
\title{Boxplots}
\usage{
\method{boxplot}{data.frame}(x, .y, .x, log, horizontal = TRUE, main = TRUE,
  crit = 1.3, ref = NULL, guide = NA_character_, ...)

\method{boxplot}{folded}(x, .y, .x, log, horizontal, main = TRUE,
  crit = 1.3, ref = NULL, ...)
}
\arguments{
\item{x}{data.frame}

\item{.y}{y axis item}

\item{.x}{x axis item}

\item{log}{whether to log transform continuous variable}

\item{horizontal}{whether box/whisker axis should be horizontal}

\item{main}{whether to include title indicating x and y items}

\item{crit}{if log is missing, log transform if mean/median ratio for non-missing x  is greater than this value}

\item{ref}{optional reference line on continuous axis}

\item{guide}{optional encoding for categories see \code{encode::encode}}

\item{...}{passed arguments}
}
\description{
Boxplots

Boxplot for data.frame.

Boxplot for folded.
}
\seealso{
Other bivariate plots: \code{\link{metaplot}},
  \code{\link{scatter}}
}
