% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{boxplot_panel}
\alias{boxplot_panel}
\title{Panel Function for Metaplot Boxplot}
\usage{
boxplot_panel(ref = NULL, horizontal, pch = "|", notch = FALSE, ref.col,
  ref.lty, ref.lwd, ref.alpha, verbose = FALSE, ...)
}
\arguments{
\item{ref}{numeric}

\item{horizontal}{passed to \code{\link[lattice]{panel.bwplot}}}

\item{pch}{passed to \code{\link[lattice]{panel.bwplot}}}

\item{notch}{passed to \code{\link[lattice]{panel.bwplot}}}

\item{ref.col}{passed to \code{\link[lattice]{panel.abline}} as col}

\item{ref.lty}{passed to \code{\link[lattice]{panel.abline}} as lty}

\item{ref.lwd}{passed to \code{\link[lattice]{panel.abline}} as lwd}

\item{ref.alpha}{passed to \code{\link[lattice]{panel.abline}} as alpha}

\item{verbose}{generate messages describing process}
}
\description{
Panel function for metaplot boxplot. Calls, \code{\link[lattice]{panel.bwplot}} and plots reference lines if ref has length.
}
\seealso{
Other panel functions: \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_gg_scatter}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_key}}, \code{\link{metaplot_ref}},
  \code{\link{panel.meta_densityplot}},
  \code{\link{panel_tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other mixedvariate plots: \code{\link{boxplot.data.frame}},
  \code{\link{boxplot_data_frame}}
}
\keyword{internal}
