% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caculate_relative_position.R
\name{calculate_relative_position}
\alias{calculate_relative_position}
\title{calculate relative position for each features}
\usage{
calculate_relative_position(
  bed_file = NULL,
  features_anno = NULL,
  scale_region = FALSE,
  cut_ratio = NULL
)
}
\arguments{
\item{bed_file}{bed format of peaks files, or a \code{GRanges}/\code{data.frame} of
peaks already loaded into R (must contain \code{seqnames}, \code{start}, \code{end}
columns).}

\item{features_anno}{features_anno object from prepare_features functions.}

\item{scale_region}{whether do scale for each feature length, default FALSE.}

\item{cut_ratio}{whether use custom ratio to plot feature regions,
for eaxample: c(0.1,0.8,0.1), default NULL.}
}
\value{
data frame
}
\description{
calculate relative position for each features
}
