% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculatePeaksRelativeDist.R
\name{CalculatePeaksRelativeDist}
\alias{CalculatePeaksRelativeDist}
\title{calculate relative position for motif}
\usage{
CalculatePeaksRelativeDist(
  peaksfa = NULL,
  motif = "[G|A][G|A]AC[A|C|T]|[T|G|A]GT[C|T][C|T]",
  mid = 3,
  pythonPath = NULL
)
}
\arguments{
\item{peaksfa}{path of fasta file for motif search.}

\item{motif}{motif sequence for searching, regular expression is accepted,
default "\link{G|A}AC\link{A|C|T}|\link{T|G|A}GT\link{C|T}" which means "RRACH".}

\item{mid}{the position for calculation the relative position, default 3.}

\item{pythonPath}{python path.}
}
\value{
the relative position.
}
\description{
calculate relative position for motif
}
