% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical.data.frame}
\alias{categorical.data.frame}
\title{Categorical Method for Data Frame}
\usage{
\method{categorical}{data.frame}(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{other arguments}
}
\value{
character
}
\description{
Categorical method for 'data.frame'.  Currently unimplemented. Returns a named vector indicating whether anonymous arguments were detected as numeric or categorical.
}
\seealso{
Other categorical: \code{\link{categorical}}

Other methods: \code{\link{axislabel.data.frame}},
  \code{\link{boxplot.data.frame}},
  \code{\link{corsplom.data.frame}},
  \code{\link{densplot.data.frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{pack.data.frame}},
  \code{\link{scatter.data.frame}},
  \code{\link{unpack.data.frame}}
}
