% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical}
\alias{categorical}
\title{Categorical Plot}
\usage{
categorical(x, ...)
}
\arguments{
\item{x}{object of dispatch}

\item{...}{passed arugments}
}
\description{
Categorical Plot. Generic, with method for 'data.frame'.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{densplot}},
  \code{\link{metaplot}}, \code{\link{pack}},
  \code{\link{scatter}}, \code{\link{test_metaplot}},
  \code{\link{unpack}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{categorical_panel}}, \code{\link{panel_tile}}
}
