% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_panel}
\alias{categorical_panel}
\title{Panel Function for Metaplot Categorical Plot}
\usage{
categorical_panel(x, y, groups, loc = getOption("metaplot_loc", 5),
  msg = getOption("metaplot_msg", "tilestats"), ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{groups}{optional grouping item}

\item{loc}{where to print statistics in a tile}

\item{msg}{a function to print text in a tile: called with x values, y values, and \dots.}

\item{...}{passed to msg}
}
\description{
Default panel function for categorical_data_frame. Implements a simple mosaic plot.
}
\seealso{
\code{\link{tilestats}}

\code{\link{categorical.data.frame}}

Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_ref}}, \code{\link{panel.tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{categorical}}, \code{\link{panel.tile}}
}
