% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{cax}
\alias{cax}
\title{Calculate Categorical Axis Labels and Positions}
\usage{
cax(x, ...)
}
\arguments{
\item{x}{x values}

\item{...}{other arguments}
}
\value{
data.frame
}
\description{
Calculates axis labels and positions for categorical values.
}
\seealso{
\code{\link{categorical_panel}}

Other categorical family: 
\code{\link{tilestats}()},
\code{\link{tiles}()}
}
\concept{categorical family}
