% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{continuous.data.frame}
\alias{continuous.data.frame}
\title{Check if Data Frame is Continuous by Item}
\usage{
\method{continuous}{data.frame}(x, col = names(x), ...)
}
\arguments{
\item{x}{data.frame}

\item{col}{a column}

\item{...}{passed arguments}
}
\description{
Checks if data.frame is continuous with respect to the column name(s).
}
\keyword{internal}
