% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{continuous.default}
\alias{continuous.default}
\title{Check if Something is Continuous by Default}
\usage{
\method{continuous}{default}(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{passed arguments}
}
\description{
Checks if something is continuous using default method.
}
\keyword{internal}
