% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{continuous.folded}
\alias{continuous.folded}
\title{Check if Folded is Continuous}
\usage{
\method{continuous}{folded}(x, var, ...)
}
\arguments{
\item{x}{folded}

\item{var}{length-one character}

\item{...}{passed arguments}
}
\description{
Checks if folded is continuous with respect to \code{var}
}
\keyword{internal}
