% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{continuous}
\alias{continuous}
\alias{continuous}
\title{Check if Something is Continuous}
\usage{
continuous(x, var, ...)

continuous(x, var, ...)
}
\arguments{
\item{x}{object}

\item{var}{length-one character}

\item{...}{passed arguments}

\item{x}{folded}

\item{...}{passed arguments}
}
\description{
Checks if something is continuous.

Checks if folded is continuous with respect to \code{var}
}
\keyword{internal}
