% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom.data.frame}
\alias{corsplom.data.frame}
\title{Correlated Scatterplot Matrix Method for Data Frame}
\usage{
\method{corsplom}{data.frame}(x, ..., fun = getOption("metaplot_corsplom",
  "corsplom_data_frame"))
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to fun}

\item{fun}{function to do the actual plotting}
}
\description{
Creates a scatterplot matrix.  Parses arguments and generates the call: fun(x, xvar, ...).
}
\seealso{
Other multivariate plots: \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.data.frame}}

Other corsplom: \code{\link{corsplom_data_frame}},
  \code{\link{corsplom}}
}
