% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom.folded}
\alias{corsplom.folded}
\title{Correlated Splom for Folded}
\usage{
\method{corsplom}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{unquoted names of variables to plot, or other named arguments}
}
\description{
Creates a scatterplot matrix with correlations for folded.
Categoricals in \dots are currently ignored. dots (\dots) are
names of items in VARIABLE to be plotted, or named arguments
to be passed to data.frame method.
}
\seealso{
Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}},
  \code{\link{overlay.data.frame}},
  \code{\link{overlay.folded}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom_data_frame}}, \code{\link{corsplom}}
}
