% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{corsplom_.folded}
\alias{corsplom_.folded}
\title{Scatterplot Matrix with Correlations for Folded, Standard Evaluation
Creates a scatterplot matrix with correlations for class 'folded'. Categoricals in \code{var} are currently ignored.}
\usage{
\method{corsplom_}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{unnamed arguments indicating variables to plot, and named arguments passed to corsplom()}
}
\description{
Scatterplot Matrix with Correlations for Folded, Standard Evaluation
Creates a scatterplot matrix with correlations for class 'folded'. Categoricals in \code{var} are currently ignored.
}
\seealso{
corsplom_.data.frame
}
