% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{corsplom_}
\alias{corsplom_}
\title{Scatterplot Matrix with Correlations, Generic for Standard Evaluation}
\usage{
corsplom_(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Generic function for scatterplot matrix with correlations.
}
\keyword{internal}
