% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom_gg_correlation}
\alias{corsplom_gg_correlation}
\title{Correlation GG Function for GG Metaplot Corsplom}
\usage{
corsplom_gg_correlation(data, mapping, col = "blue", loess.col, loess.lty,
  loess.alpha, ...)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{col}{text color}

\item{loess.col}{ignored}

\item{loess.lty}{ignored}

\item{loess.alpha}{ignored}

\item{...}{ignored}
}
\description{
Default lower panel function for corsplom_data_frame() with \code{gg = TRUE}. Plots Pearson correlation coefficient.
corsplom_data_frame() typically supplies loess aesthetics in case smooth is desired; loess arguments are defined to
suppress corresponding warnings.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_gg_scatter}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_key}}, \code{\link{metaplot_ref}},
  \code{\link{panel.meta_densityplot}},
  \code{\link{panel_tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}
}
\keyword{internal}
