% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom_gg_scatter}
\alias{corsplom_gg_scatter}
\title{Scatter GG Function for Metaplot Corsplom}
\usage{
corsplom_gg_scatter(data, mapping, method = "loess", se = F, col = "blue",
  loess.col = "blue", loess.alpha = 1, loess.lty = "solid", ...)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{col}{point color}

\item{loess.col}{loess color}

\item{loess.alpha}{loess alpha}

\item{loess.lty}{loess line type}

\item{...}{passed arguments}
}
\description{
Default upper panel function for corsplom_data_frame with \code{gg = TRUE}. Plots data with loess smooth.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_ref}}, \code{\link{panel_tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}
}
\keyword{internal}
