% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corsplom_panel_scatter}
\alias{corsplom_panel_scatter}
\title{Scatter Panel Function for Metaplot Corsplom}
\usage{
corsplom_panel_scatter(x, y, col, loess.col = getOption("metaplot_loess.col",
  col), loess.lty = getOption("metaplot_loess.lty", "solid"),
  loess.alpha = getOption("metaplot_loess.alpha", 1), ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{col}{point color}

\item{loess.col}{loess color}

\item{loess.lty}{loess line type}

\item{loess.alpha}{loess alpha}

\item{...}{passed arguments}
}
\description{
Default upper panel function for corsplom_data_frame. Plots data with loess smooth.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_ref}}, \code{\link{panel.tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}
}
\keyword{internal}
