% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dens.data.frame}
\alias{dens.data.frame}
\title{Dens Method for Data Frame}
\usage{
\method{dens}{data.frame}(x, ..., fun = getOption("metaplot_dens",
  "dens_data_frame"))
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to fun}

\item{fun}{plotting function}
}
\description{
Plot density for object of class 'data.frame'. Parses arguments and generates the call: fun(x, xvar, groups, facets,...).
}
\examples{
dens(Theoph, conc, grid = TRUE )
dens(Theoph, conc, Subject )
dens(Theoph, conc, , Subject )
}
\seealso{
Other univariate plots: \code{\link{dens_data_frame}},
  \code{\link{metaplot.data.frame}}

Other dens: \code{\link{dens_data_frame}}
}
