% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{dens.folded}
\alias{dens.folded}
\title{Plot Density for Folded}
\usage{
\method{dens}{folded}(x, var, xref = NULL, log = FALSE, ...)
}
\arguments{
\item{x}{folded}

\item{var}{item to plot}

\item{xref}{optional numeric}

\item{log}{logical; use log scale?}

\item{...}{passed arguments}
}
\description{
Plot density for object of class 'folded'.
}
