% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dens}
\alias{dens}
\title{Density Plot}
\usage{
dens(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a density plot.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{guide}},
  \code{\link{label}}, \code{\link{metaplot}},
  \code{\link{scatter}}
}
