% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dens_data_frame}
\alias{dens_data_frame}
\title{Density Function for Data Frame}
\usage{
dens_data_frame(x, var, xlab = NULL, ref = NULL, log = FALSE,
  aspect = 1, scales = NULL, panel = NULL, ...)
}
\arguments{
\item{x}{data.frame}

\item{var}{item to plot, given as length-one character}

\item{xlab}{x axis label}

\item{ref}{optional numeric}

\item{log}{whether to use log scale}

\item{aspect}{passed to \code{\link[lattice]{densityplot}}}

\item{scales}{passed to \code{\link[lattice]{densityplot}}}

\item{panel}{passed to \code{\link[lattice]{densityplot}}}

\item{...}{passed to \code{\link[lattice]{densityplot}}}
}
\description{
Plot density for object of class 'data.frame' using standard evaluation.
}
\examples{
dens_data_frame(Theoph, 'Wt', grid = TRUE)
}
\seealso{
Other univariate plots: \code{\link{dens.data.frame}},
  \code{\link{dens.folded}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}}

Other dens: \code{\link{dens.data.frame}},
  \code{\link{dens.folded}}
}
