% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densplot.data.frame}
\alias{densplot.data.frame}
\title{Densplot Method for Data Frame}
\usage{
\method{densplot}{data.frame}(x, ..., fun = getOption("metaplot_dens",
  "densplot_data_frame"))
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to fun}

\item{fun}{plotting function}
}
\description{
Plot density for object of class 'data.frame'. Parses arguments and generates the call: fun(x, xvar, groups, facets,...).
}
\examples{
densplot(Theoph, conc, grid = TRUE )
densplot(Theoph, conc, Subject )
densplot(Theoph, conc, , Subject )
}
\seealso{
Other univariate plots: \code{\link{densplot_data_frame}},
  \code{\link{metaplot.data.frame}}

Other densplot: \code{\link{densplot_data_frame}}
}
