% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densplot}
\alias{densplot}
\title{Density Plot}
\usage{
densplot(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a density plot.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{metaplot}},
  \code{\link{scatter}}
}
