% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Description_Table.R
\name{Description_Table}
\alias{Description_Table}
\title{Internal Function}
\usage{
Description_Table(source_tbl, study, group_var = NULL,
  disp_vars = NULL)
}
\arguments{
\item{source_tbl}{Dataframe containing the meta-analysis data}

\item{study}{Index column for unique study identifier e.g. "Obelix 50 B.c."}

\item{group_var}{Column to group by. Must be contained in \emph{disp_vars}.}

\item{disp_vars}{List of columns to display}
}
\value{
none
}
\description{
This function creates a six column table of the format "median ( LCI - UCI )". Called by Forest_Plot
}
