% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_pin}
\alias{diag_pin}
\title{Calculate Pin Placement}
\usage{
diag_pin(x, varname, .data, ...)
}
\arguments{
\item{x}{vector of data}

\item{varname}{name of vector in .data}

\item{.data}{original dataset, possibly with column attributes such as 'reference'}

\item{...}{passed arguments}
}
\value{
numeric
}
\description{
Calculates pin placement in the density region, inside margin of diagonal panels.
}
\seealso{
Other panel functions: \code{\link{diag_label}},
  \code{\link{l.p}}, \code{\link{metapanel_ref}},
  \code{\link{metapanel}}, \code{\link{metaplot_ref}},
  \code{\link{my.diag.panel}}, \code{\link{u.p}}

Other reference lines: \code{\link{metapanel_ref}},
  \code{\link{metaplot_ref}}
}
