% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMotif.R
\name{findMotif}
\alias{findMotif}
\title{find motif with given fasta files}
\usage{
findMotif(
  peaksfa = NULL,
  motif = "[G|A][G|A]AC[A|C|T]|[T|G|A]GT[C|T][C|T]",
  pythonPath = NULL
)
}
\arguments{
\item{peaksfa}{path of fasta file for motif search}

\item{motif}{motif sequence for searching, regular expression is accepted,
default \code{"[G|A][G|A]AC[A|C|T]|[T|G|A]GT[C|T][C|T]"} which means "RRACH".}

\item{pythonPath}{python path.}
}
\value{
target motif
}
\description{
find motif with given fasta files
}
