% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeqFromBed.R
\name{getSeqFromBed}
\alias{getSeqFromBed}
\title{extract sequence from peaks file}
\usage{
getSeqFromBed(
  peak = NULL,
  genomefie = NULL,
  outfasta = NULL,
  type = "bed",
  a = 0,
  b = 0,
  minlen = 5,
  pythonPath = NULL
)
}
\arguments{
\item{peak}{bed format file or narrow peaks file.}

\item{genomefie}{the path for genome file.}

\item{outfasta}{the path for output fasta file.}

\item{type}{input data file, "bed" or "narrowpeak", default "bed".}

\item{a}{the upstream extend bp for peak, default 0.}

\item{b}{the downstream extend bp for peak, default 0.}

\item{minlen}{the minimum peak length to extract sequence, default 5.}

\item{pythonPath}{python path.}
}
\value{
fasta file
}
\description{
extract sequence from peaks file
}
