% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guide.folded}
\alias{guide.folded}
\title{Extract Guide for Folded}
\usage{
\method{guide}{folded}(x, var, ...)
}
\arguments{
\item{x}{folded}

\item{var}{length-one character}

\item{...}{ignored arguments}
}
\value{
length-one character, possibly NA
}
\description{
Extracts guide for class folded, given a variable.
}
\keyword{internal}
