% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{guide}
\alias{guide}
\title{Extract Guide}
\usage{
guide(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts guide.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{continuous}}, \code{\link{corsplom}},
  \code{\link{dens}}, \code{\link{label}},
  \code{\link{metaplot}}, \code{\link{overlay}},
  \code{\link{scatter}}
}
\keyword{internal}
