% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{hide.folded}
\alias{hide.folded}
\title{Normalize a Folded Data Frame}
\usage{
\method{hide}{folded}(x, tolower = TRUE, ...)
}
\arguments{
\item{x}{folded}

\item{tolower}{whether to coerce attribute names to lower case}

\item{...}{other arguments}
}
\value{
data.frame
}
\description{
Convert folded data.frame to conventional format with column attributes. Scalar metadata is converted to column attributes. Other metadata left unfolded.
}
\examples{
library(fold)
data(eventsf)
head(hide(eventsf))
attributes(hide(eventsf)$BLQ)

}
\seealso{
\code{\link[fold]{fold.data.frame}}

Other hide: \code{\link{hide.data.frame}},
  \code{\link{unhide.data.frame}}
}
