% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{hide}
\alias{hide}
\title{Hide Something}
\usage{
hide(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Hide Something.  Generic, with method for data.frame.
}
\seealso{
unhide
}
