% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{indiplot.character}
\alias{indiplot.character}
\title{Make Individual Plots from Character}
\usage{
\method{indiplot}{character}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Makes individual plots from character by coercing to 'folded' (i.e. calling fold() ).
}
