% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{indiplot.numeric}
\alias{indiplot.numeric}
\title{Make Individual Plots from Numeric}
\usage{
\method{indiplot}{numeric}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Makes individual plots from numeric
}
