% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{iso_prepanel}
\alias{iso_prepanel}
\title{Prepanel Function for Isometric Axes}
\usage{
iso_prepanel(x, y, ...)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}
}
\value{
list
}
\description{
Prepanel function for isometric axes.  Returns join minimum and maximum for limits on both axes.
}
\seealso{
Other panel functions: 
\code{\link{boxplot_panel}()},
\code{\link{categorical_panel}()},
\code{\link{corsplom_gg_correlation}()},
\code{\link{corsplom_gg_diagonal}()},
\code{\link{corsplom_gg_scatter}()},
\code{\link{corsplom_panel_correlation}()},
\code{\link{corsplom_panel_diagonal}()},
\code{\link{corsplom_panel_scatter}()},
\code{\link{dens_panel}()},
\code{\link{diag_label}()},
\code{\link{diag_pin}()},
\code{\link{metaplot_key}()},
\code{\link{metaplot_ref}()},
\code{\link{panel.meta_densityplot}()},
\code{\link{panel_tile}()},
\code{\link{scatter_panel_ref}()},
\code{\link{scatter_panel}()}

Other bivariate plots: 
\code{\link{metaplot.data.frame}()},
\code{\link{scatter.data.frame}()},
\code{\link{scatter_data_frame}()},
\code{\link{scatter}()}
}
\concept{bivariate plots}
\concept{panel functions}
\keyword{internal}
