% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label.folded}
\alias{label.folded}
\title{Extract Label for Folded}
\usage{
\method{label}{folded}(x, var, ...)
}
\arguments{
\item{x}{folded}

\item{var}{length-one character}

\item{...}{ignored arguments}
}
\value{
length-one character, possibly NA
}
\description{
Extracts label for class folded, given a variable.
}
\keyword{internal}
