% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label}
\alias{label}
\title{Extract Label}
\usage{
label(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Extracts label.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{continuous}}, \code{\link{corsplom}},
  \code{\link{dens}}, \code{\link{guide}},
  \code{\link{metaplot}}, \code{\link{overlay}},
  \code{\link{scatter}}
}
\keyword{internal}
