% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forrest_plot.R
\name{make_forest_plot}
\alias{make_forest_plot}
\title{Make a forest plot}
\usage{
make_forest_plot(meta_obj)
}
\arguments{
\item{meta_obj}{A barebones meta-analysis object from psychmeta}
}
\value{
A list of plots made using ggplot
}
\description{
Create a forest plot (intentionally spelled as provided) from a meta-analytic
summary from psychmeta
}
\examples{
# make_forest_plot(meta_obj)
}
