% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_funnel_plot.R
\name{make_funnel_plot}
\alias{make_funnel_plot}
\title{Make a funnel plot}
\usage{
make_funnel_plot(
  meta_obj,
  sig_level = 0.05,
  show_sig_level_lines = TRUE,
  show_expected_dist_lines = TRUE,
  sig_label = "Region of non-significance",
  expected_label = "Expected random sampling range\\n(Assuming no moderators)",
  sig_color = "firebrick",
  expected_color = "blue",
  legend_title = "Interpretation of Lines",
  legend_position = "right"
)
}
\arguments{
\item{meta_obj}{A barebones meta-analysis object from psychmeta}

\item{sig_level}{.05}

\item{show_sig_level_lines}{TRUE}

\item{show_expected_dist_lines}{TRUE}

\item{sig_label}{"Region of non-significance}

\item{expected_label}{= "Expected random sampling range (Assuming no moderators)"}

\item{sig_color}{= "firebrick"}

\item{expected_color}{= "blue"}

\item{legend_title}{= "Interpretation of Lines"}

\item{legend_position}{= "right"}
}
\value{
A list of plots made using ggplot
}
\description{
Create a funnel plot (intentionally spelled as provided) from a meta-analytic
summary from psychmeta
}
\examples{
# make_funnel_plot(meta_obj)
}
