% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metapanel_ref}
\alias{metapanel_ref}
\title{Calculate Panel Reference Values}
\usage{
metapanel_ref(a, b, ...)
}
\arguments{
\item{a}{vector of interest}

\item{b}{vector for other axis}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Calculates reference values for x and y axes at the panel level.
}
\seealso{
Other panel functions: \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{l.p}},
  \code{\link{metapanel}}, \code{\link{metaplot_ref}},
  \code{\link{my.diag.panel}}, \code{\link{u.p}}

Other reference lines: \code{\link{diag_pin}},
  \code{\link{metaplot_ref}}
}
