\name{metaplot-internal}
\alias{metaPlot.default}
\alias{metaPlot.metaDF}
\alias{metaPlot.metaM}
\alias{meta2DF.meta.MH}
\alias{meta2DF.meta.DSL}
\alias{meta2DF.rma.mh}
\alias{meta2DF.rma.peto}
\alias{meta2DF.rma.uni}
\alias{forestDF}
\alias{forestDF.metabin}
\alias{forestDF.metacont}
\alias{forestDF.meta.MH}
\alias{forestDF.meta.DSL}
\alias{forestDF.rma.mh}
\alias{forestDF.rma.peto}
\alias{forestDF.rma.uni}


\title{Internal metaplot Functions}
\description{
  Internal functions for the package metaplot
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
