% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot.data.frame}
\alias{metaplot.data.frame}
\title{Create Metaplot from Data.frame.}
\usage{
\method{metaplot}{data.frame}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a metaplot from data.frame.  Folds x and calls the method for folded. Dots arguments are passed only to metaplot.
}
\seealso{
Other metaplot: \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}}, \code{\link{metaplot}}

Other univariate plots: \code{\link{dens.data.frame}},
  \code{\link{dens.folded}}, \code{\link{dens_data_frame}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}}

Other bivariate plots: \code{\link{boxplot.folded}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}},
  \code{\link{scatter.data.frame}},
  \code{\link{scatter.folded}},
  \code{\link{scatter_data_frame}}

Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom.folded}},
  \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}},
  \code{\link{overlay.data.frame}},
  \code{\link{overlay.folded}}
}
