% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot.folded}
\alias{metaplot.folded}
\title{Create Metaplot from Folded}
\usage{
\method{metaplot}{folded}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a plot from folded.  Packs metadata into attributes and calls next method.
}
\seealso{
Other metaplot: \code{\link{categorical}},
  \code{\link{metaplot.data.frame}}, \code{\link{metaplot}}

Other univariate plots: \code{\link{dens.data.frame}},
  \code{\link{dens.folded}}, \code{\link{dens_data_frame}},
  \code{\link{metaplot.data.frame}}

Other bivariate plots: \code{\link{boxplot.folded}},
  \code{\link{metaplot.data.frame}},
  \code{\link{scatter.data.frame}},
  \code{\link{scatter.folded}},
  \code{\link{scatter_data_frame}}

Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom.folded}},
  \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.data.frame}}
}
