% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot.grouped_df}
\alias{metaplot.grouped_df}
\title{Create Metaplot from Grouped_df}
\usage{
\method{metaplot}{grouped_df}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a metaplot from grouped_df.  Folds x and calls the method for folded. Dots arguments are passed only to metaplot.
}
\seealso{
Other metaplots: \code{\link{metapanel}}

Other univariate plots: \code{\link{dens.data.frame}},
  \code{\link{dens.folded}}, \code{\link{dens_data_frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}}

Other bivariate plots: \code{\link{boxplot.folded}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}},
  \code{\link{scatter.data.frame}},
  \code{\link{scatter.folded}},
  \code{\link{scatter_data_frame}}

Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom.folded}},
  \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}},
  \code{\link{overlay.data.frame}},
  \code{\link{overlay.folded}}

Other metaplot: \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}}, \code{\link{metaplot}}
}
