% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot_.folded}
\alias{metaplot_.folded}
\title{Metaplot Method for Folded}
\usage{
\method{metaplot_}{folded}(x, var, ...)
}
\arguments{
\item{x}{folded}

\item{var}{character: names of items to plot}

\item{...}{passed arguments}
}
\description{
Metaplot method for folded.
}
\details{
Try calling \code{\link{metaplot}} to create var from unquoted, anonymous arguments.

What happens next depends on the number and type of items represented by var.

A single argument representing a continuous variable is forwarded to \code{\link{dens}} to give a density plot.  Same for a single categorical, but this is unexpected.

Two arguments of type continuous, categorical or categorical, continuous are forwarded to \code{link{boxplot.folded}} to give a boxplot (vertical or horizontal, respectively).

Two arguments representing continuous variables give a scatterplot by means of \code{\link{scatter.folded}}.

A third anonymous argument is unexpected if a preceding argument is categorical.

A third, categorical argument following two continuous arguments is treated as a grouping variable.

If there are three or more continuous arguments, a scatterplot matrix is created by means of \code{\link{corsplom.folded}}.  Additional categoricals will be ignored.

Stratification, e.g. conditioning for trellis plots, is currently unimplemented.
}
