% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot_}
\alias{metaplot_}
\title{Metaplot Generic, Standard Evaluation}
\usage{
metaplot_(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}

\item{var}{character: quoted names of variables to plot}
}
\description{
Generic metaplot function using standard evaluation.
}
\keyword{internal}
