% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metaplot_ref}
\alias{metaplot_ref}
\title{Calculate Reference Values}
\usage{
metaplot_ref(x, var, ...)
}
\arguments{
\item{x}{data.frame}

\item{var}{name of vector in x}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Calculates reference values for x and y axes.
}
\seealso{
Other panel functions: \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{l.p}},
  \code{\link{metapanel_ref}}, \code{\link{metapanel}},
  \code{\link{my.diag.panel}}, \code{\link{u.p}}

Other reference lines: \code{\link{diag_pin}},
  \code{\link{metapanel_ref}}
}
