% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metastats}
\alias{metastats}
\title{Format GLM Statistics}
\usage{
metastats(x, y, family = if (all(y \%in\% 0:1, na.rm = TRUE)) "binomial" else
  "gaussian", ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{family}{regression family}

\item{...}{other arguments}
}
\value{
character
}
\description{
Formats GLM statistics.
}
\seealso{
\code{\link{metapanel}}

Other regression functions: \code{\link{model}},
  \code{\link{region}}
}
