% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{model}
\alias{model}
\title{Execute Linear Model}
\usage{
model(x, y, family = if (all(y \%in\% 0:1, na.rm = TRUE)) "binomial" else
  "gaussian", ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{family}{gaussian by default, or binomial for all y either zero or 1}

\item{...}{passed to \code{\link[stats]{glm}}}
}
\value{
glm
}
\description{
Executes a linear model, automatically choosing binomial family as necessary.
}
\seealso{
Other regression functions: \code{\link{metastats}},
  \code{\link{region}}
}
