% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Number_Table.R
\name{Number_Table}
\alias{Number_Table}
\title{Internal Function}
\usage{
Number_Table(
  source_tbl,
  study,
  group_var = NULL,
  median = "sens",
  lci = "sens.ci.lower",
  uci = "sens.ci.upper",
  round_digits = 2
)
}
\arguments{
\item{source_tbl}{Dataframe containing the meta-analysis data}

\item{study}{Index column for unique study identifier e.g. "Obelix 50 B.c."}

\item{group_var}{Column to group by}

\item{median}{Median}

\item{lci}{Lower boundary of confidence interval}

\item{uci}{Upper boundary of confidence interval}

\item{round_digits}{Number of decimal places to display}
}
\value{
none
}
\description{
This function creates a six column table of the format "median ( LCI - UCI )". Called by Forest_Plot
}
