% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay.R
\name{overlay}
\alias{overlay}
\title{Create an Overlay Plot}
\usage{
overlay(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates an Overlay Plot. Generic, with methods for classes 'folded' and 'data.frame'
}
\seealso{
\code{\link{overlay.folded}}

\code{\link{overlay.data.frame}}
}
