% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack.folded}
\alias{pack.folded}
\title{Normalize a Folded Data Frame}
\usage{
\method{pack}{folded}(x, tolower = TRUE, ...)
}
\arguments{
\item{x}{folded}

\item{tolower}{whether to coerce attribute names to lower case}

\item{...}{other arguments}
}
\value{
data.frame
}
\description{
Convert folded data.frame to conventional format with column attributes. Scalar metadata is converted to column attributes. Other metadata left unfolded.
}
\examples{
library(fold)
data(eventsf)
head(pack(eventsf))
attributes(pack(eventsf)$BLQ)

}
\seealso{
\code{\link[fold]{fold.data.frame}}

Other pack: \code{\link{pack.data.frame}},
  \code{\link{unpack.data.frame}}
}
