% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Pack Something}
\usage{
pack(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Pack Something.  Generic, with method for data.frame.
}
\seealso{
\code{\link{unpack}}

Other pack: \code{\link{pack.data.frame}},
  \code{\link{pack.folded}},
  \code{\link{unpack.data.frame}}
}
