% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{panel.meta_densityplot}
\alias{panel.meta_densityplot}
\title{Panel Function for Metaplot Density}
\usage{
panel.meta_densityplot(
  x,
  darg = list(n = 512),
  plot.points = "jitter",
  ref = FALSE,
  groups = NULL,
  weights = NULL,
  jitter.amount = 0.01 * diff(current.panel.limits()$ylim),
  type = "p",
  verbose = FALSE,
  ...,
  identifier = "density",
  group.number
)
}
\arguments{
\item{x}{see \code{link[lattice]{panel.densityplot}}}

\item{darg}{see \code{link[lattice]{panel.densityplot}}}

\item{plot.points}{see \code{link[lattice]{panel.densityplot}}}

\item{ref}{see \code{link[lattice]{panel.densityplot}}}

\item{groups}{see \code{link[lattice]{panel.densityplot}}}

\item{weights}{see \code{link[lattice]{panel.densityplot}}}

\item{jitter.amount}{see \code{link[lattice]{panel.densityplot}}}

\item{type}{see \code{link[lattice]{panel.densityplot}}}

\item{verbose}{generate messages describing process}

\item{...}{see \code{link[lattice]{panel.densityplot}}}

\item{identifier}{see \code{link[lattice]{panel.densityplot}}}
}
\description{
Variant of panel.densityplot that supports filled area and alpha points.
}
\seealso{
Other panel functions: 
\code{\link{boxplot_panel}()},
\code{\link{categorical_panel}()},
\code{\link{corsplom_gg_correlation}()},
\code{\link{corsplom_gg_diagonal}()},
\code{\link{corsplom_gg_scatter}()},
\code{\link{corsplom_panel_correlation}()},
\code{\link{corsplom_panel_diagonal}()},
\code{\link{corsplom_panel_scatter}()},
\code{\link{dens_panel}()},
\code{\link{diag_label}()},
\code{\link{diag_pin}()},
\code{\link{iso_prepanel}()},
\code{\link{metaplot_key}()},
\code{\link{metaplot_ref}()},
\code{\link{panel_tile}()},
\code{\link{scatter_panel_ref}()},
\code{\link{scatter_panel}()}

Other univariate plots: 
\code{\link{dens_panel}()},
\code{\link{densplot.data.frame}()},
\code{\link{densplot_data_frame}()},
\code{\link{densplot}()},
\code{\link{metaplot.data.frame}()}
}
\concept{panel functions}
\concept{univariate plots}
\keyword{internal}
