% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{panel.tile}
\alias{panel.tile}
\title{Draw a Tile}
\usage{
panel.tile(x, y, group.number, group.value, alpha, border, col, loc, msg, ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{group.number}{group number}

\item{group.value}{group value}

\item{alpha}{alpha transparency}

\item{border}{border color}

\item{col}{fill color}

\item{loc}{location for output of msg}

\item{msg}{text message as a function of x and y}
}
\description{
Draws a tile in a mosaic.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_ref}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{categorical_panel}},
  \code{\link{categorical}}
}
