% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{panel_tile}
\alias{panel_tile}
\title{Draw a Tile}
\usage{
panel_tile(x, y, subscripts, group.number, group.value, col, alpha, border, loc,
  msg, .src, cex, verbose, ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{subscripts}{subscripts}

\item{group.number}{group number}

\item{group.value}{group value}

\item{col}{fill color}

\item{alpha}{alpha transparency for fill}

\item{border}{border color}

\item{loc}{location for output of msg}

\item{msg}{ignored}

\item{.src}{data source for which subscripts give x, y, msg,  and tile limits}

\item{cex}{expansion for msg text; passed to msg}

\item{verbose}{generate messages describing process}

\item{...}{passed arguments}
}
\description{
Draws a tile in a mosaic.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_gg_scatter}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_key}}, \code{\link{metaplot_ref}},
  \code{\link{panel.meta_densityplot}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{categorical_panel}},
  \code{\link{categorical}}
}
