% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRelMotif.R
\name{plotRelMotif}
\alias{plotRelMotif}
\title{motif metaplot}
\usage{
plotRelMotif(relpos = NULL, exp = NULL, linewidth = 0.5)
}
\arguments{
\item{relpos}{a list of relative position from CalculatePeaksRelativeDist function.}

\item{exp}{sample name.}

\item{linewidth}{line width.}
}
\value{
ggplot
}
\description{
motif metaplot
}
