% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_features.R
\name{prepare_features}
\alias{prepare_features}
\title{prepare annotation features from GTF file}
\usage{
prepare_features(
  gtf_file = NULL,
  features_name = c("five_prime_utr", "CDS", "three_prime_utr")
)
}
\arguments{
\item{gtf_file}{The gtf file, gzip file is accepted.}

\item{features_name}{The feature name in gtf file,
default for ensembl database gtf file
features_name= c("five_prime_utr","CDS","three_prime_utr"), if you use
ucsc database gtf file, you should change it to c("5UTR","CDS","3UTR").}
}
\value{
GRanges object
}
\description{
prepare annotation features from GTF file
}
