\name{print.metaPar}
\alias{print.metaPar}

\title{Print Method for 'metaPar' Objects}
\description{
Print method for objects of class \code{'metaPar'}.
}
\usage{
\method{print}{metaPar}(x, ...)
}

\arguments{
  \item{x}{An object of \code{'metaPar'}}
  \item{...}{Other arguments not used.}
}

\seealso{
\code{\link{metaPar}}, \code{\link{metaParDefaults}}
}
\examples{
print(metaPar())
}
