% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{print.metaplot_gtable}
\alias{print.metaplot_gtable}
\title{Print a Metaplot GG Splom}
\usage{
\method{print}{metaplot_gtable}(x, ...)
}
\arguments{
\item{x}{metaplot_gtable}

\item{\dots}{ignored}
}
\value{
gtable
}
\description{
Prints a metaplot ggplot2-type splom. Underlying object type is 'gtable'.
Defaults for \code{link{corsplom_data_frame}} prepend class 'metaplot_gtable' to
modify plot and print functionality. Print method for 'gtable' summarizes the object.
This method invokes the plot method to give interactive behavior more like splom().
}
\seealso{
\code{\link{corsplom}}

Other methods: \code{\link{axislabel.data.frame}},
  \code{\link{boxplot.data.frame}},
  \code{\link{categorical.data.frame}},
  \code{\link{corsplom.data.frame}},
  \code{\link{densplot.data.frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{pack.data.frame}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{scatter.data.frame}},
  \code{\link{unpack.data.frame}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{corsplom}},
  \code{\link{plot.metaplot_gtable}}
}
\keyword{internal}
