% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankprobability.R
\name{rank_probabilities}
\alias{rank_probabilities}
\alias{rank_probabilities.nma.ab}
\title{Extract rank probability matrix}
\usage{
rank_probabilities(x, ...)

\method{rank_probabilities}{nma.ab}(x, ...)
}
\arguments{
\item{x}{A model object holding rank probabilities.}

\item{...}{currently not used}
}
\value{
A \code{matrix}
}
\description{
A rank probability matrix is a matrix with all entries between 0 and 1. The
rows represent treatments while the columns represent ranks.
}
