% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rank_probability_plot}
\alias{rank_probability_plot}
\alias{rank_probability_plot.matrix}
\alias{rank_probability_plot.nma.ab}
\alias{sucra_plot}
\alias{sucra_plot.matrix}
\alias{sucra_plot.nma.ab}
\title{Create rank probability or SUCRA plot}
\usage{
rank_probability_plot(x, ...)

\method{rank_probability_plot}{matrix}(x, ...)

\method{rank_probability_plot}{nma.ab}(x, ...)

sucra_plot(x, ...)

\method{sucra_plot}{matrix}(x, ...)

\method{sucra_plot}{nma.ab}(x, ...)
}
\arguments{
\item{x}{An object holding rank probabilites (matrix or model object)}

\item{...}{options (see Details)}
}
\description{
Create rank probability or SUCRA plot
}
