% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveForest.R
\name{saveForest}
\alias{saveForest}
\title{A function for convenient saveing of forest plot graphics}
\usage{
saveForest(source_plot, filename, extension = ".PDF", cm_per_row = 1,
  width = 25)
}
\arguments{
\item{source_plot}{ggplot object.}

\item{filename}{Output filename without extension.}

\item{extension}{File extension, defining the file format.}

\item{cm_per_row}{height of each row in cm.}

\item{width}{witdth of graphics in cm; e.g  in MS Word, 15.5 cm would be a standard portrait width, and 25 cm a standard landscape width}
}
\description{
This function saves forest plots as generated by Forest_Plot ussing ggsave.
}
