% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scatter_panel_ref}
\alias{scatter_panel_ref}
\title{Calculate Panel Reference Values}
\usage{
scatter_panel_ref(a, b, ...)
}
\arguments{
\item{a}{vector of interest}

\item{b}{vector for other axis}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Calculates reference values for x and y axes at the panel level.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_ref}}, \code{\link{panel_tile}},
  \code{\link{scatter_panel}}

Other reference lines: \code{\link{diag_pin}},
  \code{\link{metaplot_ref}}
}
