% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setOption}
\alias{setOption}
\title{Set or Reset Metaplot Options}
\usage{
setOption(...)
}
\arguments{
\item{...}{any metaplot options can be defined, using \code{name = value}.}
}
\value{
(invisible) character vector of option names that were set or unset
}
\description{
Sets an option value in the list \code{getOption('metaplot')}.
If invoked without named arguments, option 'metaplot' is set to NULL.
Setting an existing option moves it to the end of the list (breaks ties in \code{\link{metOption}}).
}
\examples{
example(metOption)
}
\seealso{
\code{\link{metOption}}\code{\link{options}}
}
