% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankprobability.R
\name{sucra_values}
\alias{sucra_values}
\alias{sucra_values.matrix}
\alias{sucra_values.default}
\title{Extract scra values}
\usage{
sucra_values(x, ...)

\method{sucra_values}{matrix}(x, ...)

\method{sucra_values}{default}(x, ...)
}
\arguments{
\item{x}{a \code{matrix} or a model object holding rank probabilities}

\item{...}{currently not used}
}
\description{
Extract scra values
}
\seealso{
Other extractors: \code{\link{rank_probabilities}}
}
\concept{extractors}
