% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{test_metaplot}
\alias{test_metaplot}
\title{Test Metaplot Variants}
\usage{
test_metaplot()
}
\description{
Tests metaplot variants by example.  Returns null. Use \code{example(test_metaplot)}.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{densplot}}, \code{\link{metaplot}},
  \code{\link{pack}}, \code{\link{scatter}},
  \code{\link{unpack}}

Other metaplot: \code{\link{boxplot_data_frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{densplot_data_frame}},
  \code{\link{metaplot}}, \code{\link{scatter_data_frame}}
}
