% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{tiles}
\alias{tiles}
\title{Calculate Tile Limits}
\usage{
tiles(x, ..., tex = 0.9, msg = "tilestats", verbose = FALSE)
}
\arguments{
\item{x}{a data.frame with at least columns x, y, and g, possibly f1 and f2 (facets)}

\item{...}{other arguments}

\item{tex}{tile shrinkage <= 1}

\item{msg}{a function of x and y to create a tile message}

\item{verbose}{generate messages describing process}
}
\value{
data.frame
}
\description{
Calculates limits for mosaic tiles
}
\seealso{
\code{\link{categorical_panel}}

Other categorical family: \code{\link{cax}},
  \code{\link{tilestats}}
}
