% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{tilestats}
\alias{tilestats}
\title{Format Tile Statistics}
\usage{
tilestats(x, y, ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{...}{other arguments}
}
\value{
character
}
\description{
Formats statistics for a mosaic tile.
}
\seealso{
\code{\link{categorical_panel}}
}
