% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{u.p}
\alias{u.p}
\title{Upper Panel Function}
\usage{
u.p(x, y, col, loess.col = getOption("metaplot_loess.col", col),
  loess.lty = getOption("metaplot_loess.lty", "solid"),
  loess.alpha = getOption("metaplot_loess.alpha", 1), ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{col}{point color}

\item{loess.col}{loess color}

\item{loess.lty}{loess line type}

\item{loess.alpha}{loess alpha}

\item{...}{passed arguments}
}
\description{
Upper panel function for corsplom(). Plots data with loess smooth.
}
\seealso{
Other panel functions: \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{l.p}},
  \code{\link{metapanel_ref}}, \code{\link{metapanel}},
  \code{\link{metaplot_ref}}, \code{\link{my.diag.panel}}
}
\keyword{internal}
