% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{unhide.folded}
\alias{unhide.folded}
\title{Unhide a Folded Data Frame}
\usage{
\method{unhide}{folded}(x, tolower = TRUE, ...)
}
\arguments{
\item{x}{folded}

\item{tolower}{whether to coerce attribute names to lower case}

\item{...}{other arguments}
}
\value{
data.frame
}
\description{
Convert folded data.frame to 'unhidden' format with scalar metadata as row entries.
}
\examples{
library(fold)
data(eventsf)
head(unhide(eventsf))

}
\seealso{
\code{\link[fold]{fold.data.frame}} \code{\link{hide.folded}}

Other unhide: \code{\link{unhide}}
}
