% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide.R
\name{unhide}
\alias{unhide}
\title{Unhide Something}
\usage{
unhide(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Hide Something.  Generic, with method for data.frame.
}
\seealso{
hide

Other unhide: \code{\link{unhide.folded}}
}
