% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack.folded}
\alias{unpack.folded}
\title{Unpack a Folded Data Frame}
\usage{
\method{unpack}{folded}(x, tolower = TRUE, ...)
}
\arguments{
\item{x}{folded}

\item{tolower}{whether to coerce attribute names to lower case}

\item{...}{other arguments}
}
\value{
data.frame
}
\description{
Convert folded data.frame to 'unhidden' format with scalar metadata as row entries.
}
\examples{
library(fold)
data(eventsf)
head(unpack(eventsf))

}
\seealso{
\code{\link[fold]{fold.data.frame}} \code{\link{pack.folded}}

Other unpack: \code{\link{unpack}}
}
