% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{unpack}
\alias{unpack}
\title{Unpack Something}
\usage{
unpack(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Unpack Something.  Generic, with method for data.frame.
}
\seealso{
Other pack: \code{\link{pack.data.frame}},
  \code{\link{pack}}, \code{\link{unpack.data.frame}}

Other generic functions: \code{\link{axislabel}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{densplot}}, \code{\link{metaplot}},
  \code{\link{pack}}, \code{\link{scatter}},
  \code{\link{test_metaplot}}
}
