% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wikisym2plotmath}
\alias{wikisym2plotmath}
\title{Convert Wiki Symbol to Plotmath}
\usage{
wikisym2plotmath(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
expression
}
\description{
Converts wiki symbol to plotmath.  A Wiki symbol is simple text with arbitrarily nested subscript (\code{_}) and superscript (\code{^}) groupings.  Use dot (\code{.}) to explicitly terminate a grouping, and use backslash-dot (\code{\.}) for a literal dot.  Examples: \code{V_c./F}. Trailing dots need not be supplied. Leading/trailing whitespace is removed. Tab character not allowed.
}
\examples{
wikisym2plotmath('V_c./F')
wikisym2plotmath('AUC_ss')
wikisym2plotmath('C_max_ss')
wikisym2plotmath('var^\\\\eta_j')
}
\seealso{
Other formatters: \code{\link{diag_label}}
}
