% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wikisym2plotmathOne}
\alias{wikisym2plotmathOne}
\title{Convert One Wiki Symbol to Plotmath}
\usage{
wikisym2plotmathOne(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
expression
}
\description{
Converts one wiki symbol to plotmath.  A Wiki symbol is simple text with arbitrarily nested subscript (\code{_}) and superscript (\code{^}) groupings.  Use dot (\code{.}) to explicitly terminate a grouping, and use backslash-dot (\code{\.}) for a literal dot.  Examples: \code{V_c./F}. Trailing dots need not be supplied. Leading/trailing whitespace is removed. Tab character not allowed.
}
\examples{
wikisym2plotmathOne('V_c./F')
wikisym2plotmathOne('AUC_ss')
wikisym2plotmathOne('C_max_ss')
wikisym2plotmathOne('var^eta_j')
}
\seealso{
Other formatters: \code{\link{diag_label}},
  \code{\link{wikisym2plotmath}}
}
