% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axislabel.data.frame}
\alias{axislabel.data.frame}
\title{Axis Label for Data Frame}
\usage{
\method{axislabel}{data.frame}(x, var, log = FALSE, ...)
}
\arguments{
\item{x}{data.frame}

\item{var}{item of interest}

\item{log}{whether this is for a log scale}

\item{...}{passed arguments}
}
\value{
character
}
\description{
Axis label for data.frame. Substitutes label attribute if present for column name, and puts units if present in parentheses, trailing.  Puts 'log scale' in parentheses on a new line if log is TRUE.
}
\seealso{
Other axislabel: \code{\link{axislabel}}

Other methods: \code{\link{boxplot.data.frame}},
  \code{\link{categorical.data.frame}},
  \code{\link{corsplom.data.frame}},
  \code{\link{densplot.data.frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{pack.data.frame}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}},
  \code{\link{scatter.data.frame}},
  \code{\link{unpack.data.frame}}
}
\keyword{internal}
