% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{axislabel}
\alias{axislabel}
\title{Generic Axis Label}
\usage{
axislabel(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Generic axis label, with method for 'folded'.
}
\seealso{
\code{\link{axislabel.folded}}

Other generic functions: \code{\link{continuous}},
  \code{\link{corsplom}}, \code{\link{dens}},
  \code{\link{guide}}, \code{\link{label}},
  \code{\link{metaplot}}, \code{\link{overlay}},
  \code{\link{scatter}}
}
\keyword{internal}
