% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axislabel}
\alias{axislabel}
\title{Generic Axis Label}
\usage{
axislabel(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Generic axis label, with method for 'data.frame'.
}
\seealso{
Other generic functions: \code{\link{categorical}},
  \code{\link{corsplom}}, \code{\link{densplot}},
  \code{\link{metaplot}}, \code{\link{pack}},
  \code{\link{scatter}}, \code{\link{test_metaplot}},
  \code{\link{unpack}}

Other axislabel: \code{\link{axislabel.data.frame}}
}
\keyword{internal}
