% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{boxplot.data.frame}
\alias{boxplot.data.frame}
\title{Boxplot for Data Frame}
\usage{
\method{boxplot}{data.frame}(x, .y, .x, log, horizontal = TRUE, main = TRUE,
  crit = 1.3, ref = NULL, guide = NA_character_, ...)
}
\arguments{
\item{x}{data.frame}

\item{.y}{y axis item}

\item{.x}{x axis item}

\item{log}{whether to log transform continuous variable}

\item{horizontal}{whether box/whisker axis should be horizontal}

\item{main}{whether to include title indicating x and y items}

\item{crit}{if log is missing, log transform if mean/median ratio for non-missing x  is greater than this value}

\item{ref}{optional reference line on continuous axis}

\item{guide}{optional encoding for categories see \code{encode::encode}}

\item{...}{passed arguments}
}
\description{
Boxplot for data.frame.
}
