% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{boxplot.folded}
\alias{boxplot.folded}
\title{Boxplot for Folded}
\usage{
\method{boxplot}{folded}(x, .y, .x, log, horizontal, main = TRUE,
  crit = 1.3, ref = NULL, ...)
}
\arguments{
\item{x}{folded}

\item{.y}{y axis item}

\item{.x}{x axis item}

\item{log}{whether to log transform continuous variable}

\item{horizontal}{if missing, will be set to TRUE if .x is continuous}

\item{main}{whether to include title indicating x and y items.}

\item{crit}{if log is missing, log transform if mean/median ratio for non-missing x  is greater than this value}

\item{ref}{optional reference line on continuous axis}

\item{...}{passed arguments}
}
\description{
Boxplot for folded object.
}
