% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical.data.frame}
\alias{categorical.data.frame}
\title{Categorical Method for Data Frame}
\usage{
\method{categorical}{data.frame}(x, ..., fun = metOption("categorical",
  "categorical_data_frame"),
  verbose = metOption("verbose_categorical_data_frame", FALSE))
}
\arguments{
\item{x}{data.frame}

\item{...}{other arguments}

\item{fun}{function to draw the plot}

\item{verbose}{generate messages describing process}
}
\value{
character
}
\description{
Categorical method for 'data.frame'.
}
\seealso{
Other categorical: \code{\link{categorical_data_frame}},
  \code{\link{categorical_panel}},
  \code{\link{categorical}}, \code{\link{panel_tile}}

Other methods: \code{\link{axislabel.data.frame}},
  \code{\link{boxplot.data.frame}},
  \code{\link{corsplom.data.frame}},
  \code{\link{densplot.data.frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{pack.data.frame}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}},
  \code{\link{scatter.data.frame}},
  \code{\link{unpack.data.frame}}
}
