% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{categorical}
\alias{categorical}
\title{Categorical Plot}
\usage{
categorical(x, ...)
}
\arguments{
\item{x}{data.frame}

\item{...}{other arguments}
}
\value{
character
}
\description{
Categorical plot.  Currently unimplemented. Returns a named vector indicating whether anonymous arguments were detected as numeric or categorical.
}
\seealso{
Other metaplot: \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}}, \code{\link{metaplot}}
}
