% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_data_frame}
\alias{categorical_data_frame}
\title{Categorical Function for Data Frame}
\usage{
categorical_data_frame(x, yvar = NULL, xvar, groups = NULL, facets = NULL,
  ylab = metOption("metaplot_xlab_categorical", axislabel),
  xlab = metOption("metaplot_ylab_categorical", axislabel),
  na.rm = metOption("metaplot_narm_categorical", TRUE),
  aspect = metOption("metaplot_aspect_categorical", 1),
  space = metOption("metaplot_space_categorical", "right"),
  key = metOption("metaplot_key_categorical", "metaplot_key"),
  as.table = metOption("metaplot_astable_categorical", TRUE),
  prepanel = metOption("metaplot_prepanel_categorical", function(...)
  list(xlim = 0:1, ylim = 0:1)),
  scales = metOption("metaplot_scales_categorical", NULL),
  panel = metOption("metaplot_panel_categorical", categorical_panel),
  colors = metOption("metaplot_colors_categorical", NULL),
  fill = metOption("metaplot_fill_categorical", 0.5),
  lines = metOption("metaplot_lines_categorical", TRUE),
  main = metOption("metaplot_main_categorical", NULL),
  sub = metOption("metaplot_sub_categorical", NULL),
  tex = metOption("metaplot_tex_categorical", 0.9),
  rot = metOption("metaplot_rot_categorical", c(90, 0)),
  subscripts = metOption("metaplot_subscripts_categorical", TRUE),
  par.settings = metOption("metaplot_parsettings_categorical", NULL),
  padding = metOption("metaplot_padding_categorical", 1),
  loc = metOption("metaplot_msg_loc_categorical", 5),
  msg = metOption("metaplot_msg_format_categorical", "tilestats"),
  cex = metOption("metaplot_cex_categorical", 1),
  gg = metOption("metaplot_gg_categorical", FALSE), ...)
}
\arguments{
\item{x}{data.frame}

\item{yvar}{character: y variable (optional)}

\item{xvar}{character: x variable}

\item{groups}{optional grouping variable (can be missing)}

\item{facets}{optional conditioning variables}

\item{ylab}{y axis label; can be function(x = x, var = yvar, ..)}

\item{xlab}{x axis label; can be function(x = x, var = xvar, ..)}

\item{na.rm}{whether to remove data points with one or more missing coordinates}

\item{aspect}{passed to \code{\link[lattice]{bwplot}} or ggplot; use 'fill' or NULL to calculate automatically}

\item{space}{location of key (right, left, top, bottom)}

\item{key}{list: passed to \code{\link[lattice]{xyplot}} as \code{auto.key} or to \code{\link[ggplot2]{theme}}; can be a function groups name, groups levels, fill, lines, space, gg, type ('categorical'), and \dots .  See \code{\link{metaplot_key}}.}

\item{as.table}{passed to \code{\link[lattice]{xyplot}}}

\item{prepanel}{passed to \code{\link[lattice]{xyplot}} (guessed if NULL)}

\item{scales}{passed to \code{\link[lattice]{xyplot}} (guessed if NULL)}

\item{panel}{name or definition of panel function for lattice}

\item{colors}{replacements for default colors in group order}

\item{fill}{whether to fill rectangles for each group: logical, or alpha values between 0 and 1}

\item{lines}{whether to plot borders for each group: logical, or alpha values between 0 and 1}

\item{main}{character, or a function of x, yvar, xvar, groups, facets}

\item{sub}{character, or a function of x, yvar, xvar, groups, facets}

\item{tex}{tile expansion: scale factor for reducing each tile size relative to full size (<= 1)}

\item{rot}{rotation for axis labels; can be length 2 for y and x axes, respectively}

\item{subscripts}{passed to \code{\link[lattice]{xyplot}}}

\item{par.settings}{passed to \code{\link[lattice]{xyplot}} (calculated if null)}

\item{padding}{if true (and par.settings is not supplied), lattice padding will be tweaked to mimic simple ggplot layout}

\item{loc}{where to print statistics in a tile}

\item{msg}{a function of x and y to print text in a tile}

\item{cex}{expansion for msg text}

\item{gg}{logical: whether to generate \code{ggplot} instead of \code{trellis}}

\item{...}{passed to \code{\link{region}}}
}
\description{
Categorical function for class 'data.frame'. Implements a simple mosaic plot.
}
\examples{

library(magrittr)
library(dplyr)
library(csv)
x <- as.csv(system.file(package = 'metaplot', 'extdata/theoph.csv'))
x \%<>\% pack
x \%>\% metaplot(site)
x \%>\% metaplot(site, gg = T)
x \%>\% metaplot(arm, site)
x \%>\% metaplot(arm, site, gg = T)
x \%>\% metaplot(arm, site, cohort)
x \%>\% metaplot(arm, site, cohort, gg = T)
x \%>\% metaplot(arm, site, cohort, space = 'top')
x \%>\% metaplot(arm, site, , cohort)
x \%>\% metaplot(arm, site, , cohort, gg = T)
x \%>\% metaplot(arm, site, , cohort, rot = c(0,90))
x \%>\% metaplot(arm, site, , cohort, rot = c(0,90), gg = T)
x \%>\% metaplot(arm, site, , cohort, rot = c(45, 45))
x \%>\% metaplot(subject,cohort,arm, site, lines = F, rot = c(45,45))
x \%>\% metaplot(subject,cohort,arm, site, lines = F, rot = c(45,45), gg=T)
# panel-specific axis not well-supported for gg version
x \%>\% metaplot(subject,cohort,,arm, site)
x \%>\% metaplot(subject,cohort,,arm, site, gg=T)
}
\seealso{
\code{\link{categorical_panel}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_panel}},
  \code{\link{categorical}}, \code{\link{panel_tile}}

Other metaplot: \code{\link{boxplot_data_frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{densplot_data_frame}},
  \code{\link{metaplot_key}}, \code{\link{metaplot}},
  \code{\link{scatter_data_frame}},
  \code{\link{test_metaplot}}
}
