% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{categorical_data_frame}
\alias{categorical_data_frame}
\title{Categorical Function for Data Frame}
\usage{
categorical_data_frame(x, yvar = NULL, xvar, groups = NULL, facets = NULL,
  ylab = getOption("metaplot_categorical_xlab", axislabel),
  xlab = getOption("metaplot_categorical_ylab", axislabel),
  na.rm = getOption("metaplot_categorical_na_rm", TRUE),
  aspect = getOption("metaplot_categorical_aspect", 1),
  space = getOption("metaplot_categorical_space", "right"),
  key = getOption("metaplot_categorical_key", "metaplot_key"),
  as.table = getOption("metaplot_categorical_as_table", TRUE),
  prepanel = getOption("metaplot_categorical_prepanel", function(...)
  list(xlim = 0:1, ylim = 0:1)),
  scales = getOption("metaplot_categorical_scales", NULL),
  panel = getOption("metaplot_categorical_panel", categorical_panel),
  colors = getOption("metaplot_categorical_colors", NULL),
  fill = getOption("metaplot_categorical_fill", 0.5),
  lines = getOption("metaplot_categorical_lines", TRUE),
  main = getOption("metaplot_categorical_main", NULL),
  sub = getOption("metaplot_categorical_sub", NULL),
  tex = getOption("metaplot_categorical_tex", 0.9),
  rot = getOption("metaplot_categorical_rot", c(90, 0)),
  subscripts = getOption("metaplot_categorical_subscripts", TRUE),
  par.settings = getOption("metaplot_categorical_par_settings", NULL),
  padding = getOption("metaplot_categorical_padding", 1),
  loc = getOption("metaplot_categorical_loc", 5),
  msg = getOption("metaplot_categorical_msg", "tilestats"),
  cex = getOption("metaplot_categorical_panel_cex", 1),
  gg = getOption("metaplot_categorical_gg", FALSE), ...)
}
\arguments{
\item{x}{data.frame}

\item{yvar}{character: y variable (optional)}

\item{xvar}{character: x variable}

\item{groups}{optional grouping variable (can be missing)}

\item{facets}{optional conditioning variables}

\item{ylab}{y axis label; can be function(x = x, var = yvar, ..)}

\item{xlab}{x axis label; can be function(x = x, var = xvar, ..)}

\item{na.rm}{whether to remove data points with one or more missing coordinates}

\item{aspect}{passed to \code{\link[lattice]{xyplot}}}

\item{space}{location of key (right, left, top, bottom)}

\item{key}{list: passed to \code{\link[lattice]{xyplot}} as \code{auto.key} or to \code{\link[ggplot2]{theme}}; can be a function groups name, groups levels, fill, lines, space, gg, type ('categorical'), and \dots .  See \code{\link{metaplot_key}}.}

\item{as.table}{passed to \code{\link[lattice]{xyplot}}}

\item{prepanel}{passed to \code{\link[lattice]{xyplot}} (guessed if NULL)}

\item{scales}{passed to \code{\link[lattice]{xyplot}} (guessed if NULL)}

\item{panel}{name or definition of panel function for lattice}

\item{colors}{replacements for default colors in group order}

\item{fill}{whether to fill rectangles for each group: logical, or alpha values between 0 and 1}

\item{lines}{whether to plot borders for each group: logical, or alpha values between 0 and 1}

\item{main}{character, or a function of x, yvar, xvar, groups, facets}

\item{sub}{character, or a function of x, yvar, xvar, groups, facets}

\item{tex}{tile expansion: scale factor for reducing each tile size relative to full size (<= 1)}

\item{rot}{rotation for axis labels; can be length 2 for y and x axes, respectively}

\item{subscripts}{passed to \code{\link[lattice]{xyplot}}}

\item{par.settings}{passed to \code{\link[lattice]{xyplot}} (calculated if null)}

\item{padding}{if true (and par.settings is not supplied), lattice padding will be tweaked to mimic simple ggplot layout}

\item{loc}{where to print statistics in a tile}

\item{msg}{a function of x and y to print text in a tile}

\item{cex}{expansion for msg text}

\item{gg}{logical: whether to generate \code{ggplot} instead of \code{trellis}}

\item{...}{passed to \code{\link{region}}}
}
\description{
Categorical function for class 'data.frame'. Implements a simple mosaic plot.
}
\examples{

library(magrittr)
library(dplyr)
library(csv)
x <- as.csv(system.file(package = 'metaplot', 'extdata/theoph.csv'))
x \%<>\% pack
x \%>\% metaplot(site)
x \%>\% metaplot(site, gg = T)
x \%>\% metaplot(arm, site)
x \%>\% metaplot(arm, site, gg = T)
x \%>\% metaplot(arm, site, cohort)
x \%>\% metaplot(arm, site, cohort, gg = T)
x \%>\% metaplot(arm, site, cohort, space = 'top')
x \%>\% metaplot(arm, site, , cohort)
x \%>\% metaplot(arm, site, , cohort, gg = T)
x \%>\% metaplot(arm, site, , cohort, rot = c(0,90))
x \%>\% metaplot(arm, site, , cohort, rot = c(0,90), gg = T)
x \%>\% metaplot(arm, site, , cohort, rot = c(45, 45))
x \%>\% metaplot(subject,cohort,arm, site, lines = F, rot = c(45,45))
x \%>\% metaplot(subject,cohort,arm, site, lines = F, rot = c(45,45), gg=T)
# panel-specific axis not well-supported for gg version
x \%>\% metaplot(subject,cohort,,arm, site)
x \%>\% metaplot(subject,cohort,,arm, site, gg=T)
}
\seealso{
\code{\link{categorical_panel}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_panel}},
  \code{\link{categorical}}, \code{\link{panel_tile}}

Other metaplot: \code{\link{boxplot_data_frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{densplot_data_frame}},
  \code{\link{metaplot_key}}, \code{\link{metaplot}},
  \code{\link{scatter_data_frame}},
  \code{\link{test_metaplot}}
}
