% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{continuous}
\alias{continuous}
\title{Check if Something is Continuous}
\usage{
continuous(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Checks if something is continuous.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{dens}},
  \code{\link{guide}}, \code{\link{label}},
  \code{\link{metaplot}}, \code{\link{scatter}}
}
\keyword{internal}
