% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{corsplom.data.frame}
\alias{corsplom.data.frame}
\title{Scatterplot Matrix with Correlations for data.frame, Standard Evaluation
Creates a scatterplot matrix with correlations in lower panel, by default.}
\usage{
\method{corsplom}{data.frame}(x, upper.panel = u.p, lower.panel = l.p,
  pscales = 0, xlab = "", varname.cex = 0.5, diag.panel = my.diag.panel,
  split = TRUE, ...)
}
\arguments{
\item{x}{data.frame}

\item{upper.panel}{passed to splom}

\item{lower.panel}{passed to splom}

\item{pscales}{passed to splom}

\item{xlab}{passed to splom}

\item{varname.cex}{passed to splom}

\item{diag.panel}{passed to splom}

\item{split}{break diagonal names on white space}

\item{...}{passed arguments}
}
\description{
Scatterplot Matrix with Correlations for data.frame, Standard Evaluation
Creates a scatterplot matrix with correlations in lower panel, by default.
}
