% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom.data.frame}
\alias{corsplom.data.frame}
\title{Correlated Scatterplot Matrix Method for Data Frame}
\usage{
\method{corsplom}{data.frame}(x, ..., upper.panel = u.p, lower.panel = l.p,
  pscales = 0, xlab = "", varname.cex = 1, diag.panel = my.diag.panel,
  split = TRUE, fun = getOption("metaplot_corsplom", "corsplom_data_frame"))
}
\arguments{
\item{x}{data.frame}

\item{...}{variables to plot as unquoted character strings}

\item{upper.panel}{passed to splom}

\item{lower.panel}{passed to splom}

\item{pscales}{passed to splom}

\item{xlab}{passed to splom}

\item{varname.cex}{passed to splom}

\item{diag.panel}{passed to splom}

\item{split}{break diagonal names on white space}

\item{fun}{function to do the actual plotting}
}
\description{
Creates a scatterplot matrix using nonstandard evaluation.
}
\seealso{
Other multivariate plots: \code{\link{corsplom.folded}},
  \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}},
  \code{\link{overlay.data.frame}},
  \code{\link{overlay.folded}}

Other corsplom: \code{\link{corsplom.folded}},
  \code{\link{corsplom_data_frame}}, \code{\link{corsplom}}
}
