% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom.data.frame}
\alias{corsplom.data.frame}
\title{Correlated Scatterplot Matrix Method for Data Frame}
\usage{
\method{corsplom}{data.frame}(x, ..., fun = getOption("metaplot_corsplom",
  "corsplom_data_frame"))
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to fun}

\item{fun}{function to do the actual plotting}
}
\description{
Creates a scatterplot matrix.  Parses arguments and generates the call: fun(x, xvar, ...).
}
\seealso{
Other multivariate plots: \code{\link{corsplom_data_frame}},
  \code{\link{metaplot.data.frame}}

Other corsplom: \code{\link{corsplom_data_frame}},
  \code{\link{corsplom}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}}

Other methods: \code{\link{axislabel.data.frame}},
  \code{\link{boxplot.data.frame}},
  \code{\link{categorical.data.frame}},
  \code{\link{densplot.data.frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{pack.data.frame}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}},
  \code{\link{scatter.data.frame}},
  \code{\link{unpack.data.frame}}
}
