% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{corsplom.folded}
\alias{corsplom.folded}
\title{Scatterplot Matrix with Correlations for Folded, Non Standard Evaluation
Creates a scatterplot matrix with correlations for class 'folded'. Categoricals in \code{var} are currently ignored.}
\usage{
\method{corsplom}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{passed arguments}
}
\description{
Scatterplot Matrix with Correlations for Folded, Non Standard Evaluation
Creates a scatterplot matrix with correlations for class 'folded'. Categoricals in \code{var} are currently ignored.
}
\seealso{
\code{\link{corsplom_.folded}} \code{\link{corsplom_.data.frame}}
}
