% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom}
\alias{corsplom}
\title{Correlated Splom}
\usage{
corsplom(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Scatterplot matrix with correlations.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{categorical}}, \code{\link{densplot}},
  \code{\link{metaplot}}, \code{\link{pack}},
  \code{\link{scatter}}, \code{\link{test_metaplot}},
  \code{\link{unpack}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}}
}
