% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom_data_frame}
\alias{corsplom_data_frame}
\title{Correlated Scatterplot Matrix Function for Data Frame}
\usage{
corsplom_data_frame(x, xvar = names(x),
  upper.panel = getOption("metaplot_upper.panel", if (gg) corsplom_gg_scatter
  else corsplom_panel_scatter),
  lower.panel = getOption("metaplot_lower.panel", if (gg)
  corsplom_gg_correlation else corsplom_panel_correlation),
  diag.panel = getOption("metaplot_diag.panel", if (gg) corsplom_gg_diagonal
  else corsplom_panel_diagonal), pscales = getOption("metaplot_pscales", 0),
  xlab = getOption("metaplot_corsplom_xlab", NULL),
  varname.cex = getOption("metaplot_varname.cex", 1),
  main = getOption("metaplot_main", NULL), sub = getOption("metaplot_sub",
  NULL), col = getOption("metaplot_corsplom_point_col", "blue"),
  loess.col = getOption("metaplot_loess.col", col),
  loess.lty = getOption("metaplot_loess.lty", "solid"),
  loess.alpha = getOption("metaplot_loess.alpha", 1), density = TRUE,
  diag.label = getOption("metaplot_diag.label", diag_label),
  pin = getOption("metaplot_pin", diag_pin),
  pin.col = getOption("metaplot_pin.col", "darkgrey"),
  pin.alpha = getOption("metaplot_pin.alpha", 1),
  dens.col = getOption("metaplot_dens.col", "grey"),
  dens.scale = getOption("metaplot_dens.scale", 0.2),
  dens.alpha = getOption("metaplot_dens.alpha", 0.5),
  gg = getOption("metaplot_gg", FALSE), ...)
}
\arguments{
\item{x}{data.frame}

\item{xvar}{variables to plot}

\item{upper.panel}{passed to \code{\link[lattice]{splom}} or (if gg=T) \code{\link[GGally]{ggpairs}} (as 'upper')}

\item{lower.panel}{passed to \code{\link[lattice]{splom}} or (if gg=T) \code{\link[GGally]{ggpairs}} (as 'lower')}

\item{diag.panel}{passed to \code{\link[lattice]{splom}} or (if gg=T) \code{\link[GGally]{ggpairs}} (as 'diagonal')}

\item{pscales}{passed to \code{\link[lattice]{splom}}}

\item{xlab}{can be function(x = x, var = xvar, ...)}

\item{varname.cex}{text size multiplier}

\item{main}{character, or a function of x, xvar}

\item{sub}{character, or a function of x, xvar}

\item{col}{point color}

\item{loess.col}{loess color}

\item{loess.lty}{loess line type}

\item{loess.alpha}{loess alpha}

\item{density}{whether to plot density polygons}

\item{diag.label}{label for the diagonal; can be a function of x, varname, .data}

\item{pin}{location for a pin (reference line) in the density region; can be a function of x, varname, .data}

\item{pin.col}{color of pin, if any}

\item{pin.alpha}{alpha transparency of pin}

\item{dens.col}{color for density region}

\item{dens.scale}{inflation factor for height of density smooth}

\item{dens.alpha}{alpha transparency for density region}

\item{gg}{logical: whether to generate \code{ggplot} instead of \code{trellis}}

\item{...}{extra arguments passed to \code{\link[lattice]{splom}} and \code{\link[GGally]{ggpairs}} (including upper,lower, and diag)}
}
\description{
Creates a scatterplot matrix with correlations in lower panel, by default.
}
\examples{

library(magrittr)
library(dplyr)
library(csv)
x <- as.csv(system.file(package = 'metaplot', 'extdata/theoph.csv'))
x \%<>\% pack
# options(metaplot_gg = TRUE)
x \%>\% metaplot(lKe, lKa, lCl)
x \%>\% metaplot(
  lKe, lKa, lCl,
  col = 'black',loess.col = 'red', pin.col = 'red',
  dens.col='blue',dens.alpha = 0.1
)
}
\seealso{
Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{metaplot.data.frame}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom}}

Other metaplot: \code{\link{boxplot_data_frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{densplot_data_frame}},
  \code{\link{metaplot}}, \code{\link{scatter_data_frame}},
  \code{\link{test_metaplot}}
}
