% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom_data_frame}
\alias{corsplom_data_frame}
\title{Correlated Scatterplot Matrix Function for Data Frame}
\usage{
corsplom_data_frame(x, var = names(x), upper.panel = u.p,
  lower.panel = l.p, pscales = 0, xlab = "", varname.cex = 1,
  diag.panel = my.diag.panel, split = TRUE, sub = attr(x, "source"), ...)
}
\arguments{
\item{x}{data.frame}

\item{var}{variables to plot}

\item{upper.panel}{passed to splom}

\item{lower.panel}{passed to splom}

\item{pscales}{passed to splom}

\item{xlab}{passed to splom}

\item{varname.cex}{passed to splom}

\item{diag.panel}{passed to splom}

\item{split}{break diagonal names on white space}

\item{sub}{passed to splom}

\item{...}{extra arguments passed to \code{\link[lattice]{splom}}}
}
\description{
Creates a scatterplot matrix with correlations in lower panel, by default.
}
\seealso{
Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{metaplot.data.frame}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom}}
}
