% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom_data_frame}
\alias{corsplom_data_frame}
\title{Correlated Scatterplot Matrix Function for Data Frame}
\usage{
corsplom_data_frame(x, xvar = names(x),
  upper.panel = getOption("metaplot_corsplom_upper_panel", if (gg)
  corsplom_gg_scatter else corsplom_panel_scatter),
  lower.panel = getOption("metaplot_corsplom_lower_panel", if (gg)
  corsplom_gg_correlation else corsplom_panel_correlation),
  diag.panel = getOption("metaplot_corsplom_diag_panel", if (gg)
  corsplom_gg_diagonal else corsplom_panel_diagonal),
  pscales = getOption("metaplot_corsplom_pscales", 0),
  xlab = getOption("metaplot_corsplom_xlab", NULL),
  varname.cex = getOption("metaplot_corsplom_varname_cex", 1),
  main = getOption("metaplot_corsplom_main", NULL),
  sub = getOption("metaplot_corsplom_sub", NULL),
  col = getOption("metaplot_corsplom_point_col", "blue"),
  loess.col = getOption("metaplot_corsplom_loess_col", col),
  loess.lty = getOption("metaplot_corsplom_loess_lty", "solid"),
  loess.alpha = getOption("metaplot_corsplom_loess_alpha", 1),
  density = getOption("metaplot_corsplom_density", TRUE),
  diag.label = getOption("metaplot_corsplom_diag_label", diag_label),
  pin = getOption("metaplot_corsplom_pin", diag_pin),
  pin.col = getOption("metaplot_corsplom_pin_col", "darkgrey"),
  pin.alpha = getOption("metaplot_corsplom_pin_alpha", 1),
  dens.col = getOption("metaplotcorsplom_dens_col", "grey"),
  dens.scale = getOption("metaplot_corsplom_dens_scale", 0.2),
  dens.alpha = getOption("metaplot_corsplom_dens_alpha", 0.5),
  par.settings = getOption("metaplot_corsplom_par_settings", NULL),
  padding = getOption("metaplot_corsplom_padding", 1),
  as.table = getOption("metaplot_corsplom_as_table", FALSE),
  upper = getOption("metaplot_corsplom_upper", TRUE),
  gg = getOption("metaplot_corsplom_gg", FALSE), ...)
}
\arguments{
\item{x}{data.frame}

\item{xvar}{variables to plot}

\item{upper.panel}{passed to \code{\link[lattice]{splom}} or ggplot}

\item{lower.panel}{passed to \code{\link[lattice]{splom}} or ggplot}

\item{diag.panel}{passed to \code{\link[lattice]{splom}} or ggplot}

\item{pscales}{passed to \code{\link[lattice]{splom}}}

\item{xlab}{can be function(x = x, var = xvar, ...)}

\item{varname.cex}{text size multiplier}

\item{main}{character, or a function of x, xvar}

\item{sub}{character, or a function of x, xvar}

\item{col}{point color}

\item{loess.col}{loess color}

\item{loess.lty}{loess line type}

\item{loess.alpha}{loess alpha}

\item{density}{whether to plot density polygons}

\item{diag.label}{label for the diagonal; can be a function of x, varname, .data}

\item{pin}{location for a pin (reference line) in the density region; can be a function of x, varname, .data}

\item{pin.col}{color of pin, if any}

\item{pin.alpha}{alpha transparency of pin}

\item{dens.col}{color for density region}

\item{dens.scale}{inflation factor for height of density smooth}

\item{dens.alpha}{alpha transparency for density region}

\item{par.settings}{passed to \code{\link[lattice]{xyplot}} (calculated if NULL)}

\item{padding}{numeric (will be recycled to length 4) giving plot margins in default units: top, right, bottom, left (in multiples of 5.5 points for ggplot)}

\item{as.table}{diagonal arranged top-left to bottom-right}

\item{upper}{whether density plots in diagonal should face the upper triangle vs. lower}

\item{gg}{logical: whether to generate \code{ggplot} instead of \code{trellis}}

\item{...}{extra arguments passed to \code{\link[lattice]{splom}} and ggplot}
}
\value{
trellis or grob
}
\description{
Creates a scatterplot matrix with correlations in lower panel, by default.
}
\examples{

library(magrittr)
library(dplyr)
library(csv)
x <- as.csv(system.file(package = 'metaplot', 'extdata/theoph.csv'))
x \%<>\% pack
# options(metaplot_gg = TRUE)
x \%>\% metaplot(lKe, lKa, lCl)
x \%>\% metaplot(
  lKe, lKa, lCl,
  col = 'black',loess.col = 'red', pin.col = 'red',
  dens.col='blue',dens.alpha = 0.1
)
}
\seealso{
Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{metaplot.data.frame}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}}

Other metaplot: \code{\link{boxplot_data_frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{densplot_data_frame}},
  \code{\link{metaplot_key}}, \code{\link{metaplot}},
  \code{\link{scatter_data_frame}},
  \code{\link{test_metaplot}}
}
