% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corsplom.R
\name{corsplom_gg_scatter}
\alias{corsplom_gg_scatter}
\title{Scatter GG Function for Metaplot Corsplom}
\usage{
corsplom_gg_scatter(data, mapping, method = "loess", se = F,
  col = metOption("metaplot_point_col_corsplom_gg", "blue"),
  smooth.col = metOption("metaplot_smooth_col_corsplom_gg", "blue"),
  smooth.alpha = metOption("metaplot_smooth_alpha_corsplom_gg", 1),
  smooth.lty = metOption("metaplot_smooth_lty_corsplom_gg", "solid"),
  smooth.lwd = metOption("metaplot_smooth_lwd_corsplom_gg", 1), ...)
}
\arguments{
\item{data}{data}

\item{mapping}{mapping}

\item{col}{point color}

\item{smooth.col}{smooth color}

\item{smooth.alpha}{smooth alpha}

\item{smooth.lty}{smooth line type}

\item{smooth.lwd}{smooth line size}

\item{...}{passed arguments}
}
\description{
Default upper panel function for corsplom_data_frame with \code{gg = TRUE}. Plots data with loess smooth.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_key}}, \code{\link{metaplot_ref}},
  \code{\link{panel.meta_densityplot}},
  \code{\link{panel_tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other corsplom: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom_data_frame}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{corsplom}},
  \code{\link{plot.metaplot_gtable}},
  \code{\link{print.metaplot_gtable}}
}
\keyword{internal}
