% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{dens.data.frame}
\alias{dens.data.frame}
\title{Plot Density for Data Frame}
\usage{
\method{dens}{data.frame}(x, var, ref = NULL, log = FALSE, ...)
}
\arguments{
\item{x}{folded}

\item{var}{item to plot}

\item{ref}{optional numeric}

\item{log}{logical; use log scale?}

\item{...}{passed arguments}
}
\description{
Plot density for object of class 'data.frame'.
}
