% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{dens.folded}
\alias{dens.folded}
\title{Dens Method for Folded}
\usage{
\method{dens}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{passed to \code{\link{boxplot.data.frame}}}
}
\description{
Dens method for folded. Converts to data.frame with defined column attributes and calls data.frame method.
}
\examples{
library(fold)
data(eventsf)
dens(eventsf, DV)
}
\seealso{
Other univariate plots: \code{\link{dens.data.frame}},
  \code{\link{dens_data_frame}},
  \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}}

Other dens: \code{\link{dens.data.frame}},
  \code{\link{dens_data_frame}}
}
