% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{densplot}
\alias{densplot}
\title{Density Plot}
\usage{
densplot(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a density plot.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{metaplot}}, \code{\link{pack}},
  \code{\link{scatter}}, \code{\link{test_metaplot}},
  \code{\link{unpack}}

Other univariate plots: \code{\link{dens_panel}},
  \code{\link{densplot.data.frame}},
  \code{\link{densplot_data_frame}},
  \code{\link{metaplot.data.frame}}

Other densplot: \code{\link{densplot.data.frame}},
  \code{\link{densplot_data_frame}}
}
