% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_label}
\alias{diag_label}
\title{Format a Diagonal Label}
\usage{
diag_label(varname, .data,
  diag_label_simple = getOption("metaplot_diag_label_simple", FALSE),
  diag_label_split = getOption("metaplot_diag_label_split", TRUE),
  diag_symbol_format = getOption("metaplot_diag_symbol_format",
  "wikisym2plotmath"), ...)
}
\arguments{
\item{varname}{character}

\item{.data}{data.frame}

\item{diag_label_simple}{logical: just return varname?}

\item{diag_label_split}{whether to substitute line breaks for spaces}

\item{diag_symbol_format}{function to process symbol attribute, if present}

\item{...}{ignored}
}
\value{
character
}
\description{
Formats a diagonal label. Can return a simple column name, a column label (if attribute defined), a fractured column label (split on spaces), or a processed symbol (over-rides label).
}
\seealso{
Other panel functions: \code{\link{diag_pin}},
  \code{\link{l.p}}, \code{\link{metapanel_ref}},
  \code{\link{metapanel}}, \code{\link{metaplot_ref}},
  \code{\link{my.diag.panel}}, \code{\link{u.p}}

Other formatters: \code{\link{wikisym2plotmath}}
}
