% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_pin}
\alias{diag_pin}
\title{Calculate Pin Placement}
\usage{
diag_pin(x, varname, .data, ...)
}
\arguments{
\item{x}{vector of data}

\item{varname}{name of vector in .data}

\item{.data}{original dataset, possibly with column attributes such as 'reference'}

\item{...}{passed arguments}
}
\value{
numeric
}
\description{
Calculates pin placement in the density region, inside margin of diagonal panels.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{iso_prepanel}}, \code{\link{metaplot_ref}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other reference lines: \code{\link{metaplot_ref}},
  \code{\link{scatter_panel_ref}}
}
