% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{l.p}
\alias{l.p}
\title{Lower Panel Function}
\usage{
l.p(x, y, ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{...}{passed arguments}
}
\description{
Lower panel function for corsplom(). Plots Pearson correlation coefficient.
}
\seealso{
Other panel functions: \code{\link{metapanel}},
  \code{\link{my.diag.panel}}, \code{\link{u.p}}
}
\keyword{internal}
