\name{metaplot-internal}
\alias{metaPlot.default}
\alias{metaPlot.metaDF}
\alias{metaPlot.metaM}

\title{Internal metaplot Functions}
\description{
  Internal functions for the package metaplot
}
\details{
  Functions which are either not intended to be called by the user or
  are waiting to be documented.
}
\keyword{ internal }
