% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot.data.frame}
\alias{metaplot.data.frame}
\title{Create Metaplot for Data Frame.}
\usage{
\method{metaplot}{data.frame}(x, ...,
  univariate = getOption("metaplot_univariate", "densplot"),
  mixedvariate = getOption("metaplot_mixedvariate", "boxplot"),
  bivariate = getOption("metaplot_bivariate", "scatter"),
  multivariate = getOption("metaplot_multivariate", "corsplom"),
  categorical = getOption("metaplot_categorical", "categorical"))
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}

\item{univariate}{function for univariate arguments}

\item{mixedvariate}{function for bivariate combinations of numeric and categoral arguments}

\item{bivariate}{function for arguments that resolve to two numerics (see rules)}

\item{multivariate}{function for more than two numeric arguments}

\item{categorical}{function for categorical arguments}
}
\description{
Create Metaplot for Data Frame.
}
\examples{
\dontrun{
library(magrittr)
library(dplyr)
library(csv)
library(nlme)
x <- Theoph

# mixed effects model
m1 <- nlme(
  conc ~ SSfol(Dose, Time, lKe, lKa, lCl),
  data = x,
  fixed = lKe + lKa + lCl ~ 1,
  random = lKe + lKa + lCl ~ 1
)

# some numeric and categorical properties
x \%<>\% mutate(arm = ifelse(as.numeric(as.character(Subject)) \%\% 2 == 0, 1, 2))
x \%<>\% mutate(site = ifelse(as.numeric(as.character(Subject)) < 7, 1, 2))
x \%<>\% mutate(pred = predict(m1,level = 0) \%>\% signif(4))
x \%<>\% mutate(ipred = predict(m1) \%>\% signif(4))
x \%<>\% mutate(res = residuals(m1) \%>\% signif(4))
x \%<>\% mutate(sres = residuals(m1, type = 'pearson') \%>\% signif(4))
r <- ranef(m1) \%>\% signif(4)
r$Subject <- rownames(r)
x \%<>\% left_join(r)
# metadata
attr(x$Subject,'label') <- 'subject identifier'
attr(x$Wt,'label') <- 'subject weight'
attr(x$Dose,'label') <- 'theophylline dose'
attr(x$Time,'label') <- 'time since dose administration'
attr(x$conc,'label') <- 'theophylline concentration'
attr(x$arm,'label') <- 'trial arm'
attr(x$site,'label') <- 'investigational site'
attr(x$pred,'label') <- 'population-predicted concentration'
attr(x$ipred,'label') <- 'individual-predicted concentration'
attr(x$res,'label') <- 'residuals'
attr(x$sres,'label') <- 'standardized residuals'
attr(x$lKe,'label') <- 'natural log of elimination rate constant'
attr(x$lKa,'label') <- 'natural log of absorption rate constant'
attr(x$lCl,'label') <- 'natural log of clearance'
attr(x$Subject,'guide') <- '....'
attr(x$Wt,'guide') <- 'kg'
attr(x$Dose,'guide') <- 'mg/kg'
attr(x$Time,'guide') <- 'h'
attr(x$conc,'guide') <- 'mg/L'
attr(x$arm,'guide') <- '//1/Arm A//2/Arm B//'
attr(x$site,'guide') <- '//1/Site 1//2/Site 2//'
attr(x$pred,'guide') <- 'mg/L'
attr(x$ipred,'guide') <- 'mg/L'

attr(x$lKe,'reference') <- 0
attr(x$lKa,'reference') <- 0
attr(x$lCl,'reference') <- 0
attr(x$res,'reference') <- 0
attr(x$sres,'reference') <- '//-1.96//1.96//'

attr(x$Subject,'symbol') <- 'ID_i'
attr(x$Wt,'symbol') <- 'W_i'
attr(x$Dose,'symbol') <- 'A_i'
attr(x$Time,'symbol') <- 't_i,j'
attr(x$conc,'symbol') <- 'C_i,j'
attr(x$arm,'symbol') <- 'Arm_i'
attr(x$site,'symbol') <- 'Site_i'
attr(x$pred,'symbol') <- 'C_pred_p'
attr(x$ipred,'symbol') <- 'C_pred_i'
attr(x$res,'symbol') <- '\\\\epsilon'
attr(x$sres,'symbol') <- '\\\\epsilon_st'
attr(x$lKe,'symbol') <- 'ln(K_e)'
attr(x$lKa,'symbol') <- 'ln(K_a)'
attr(x$lCl,'symbol') <- 'ln(Cl_c./F'


x \%>\% unpack \%>\% as.csv('theoph.csv')
}
\dontshow{
\dontrun{
y <- x
y[] <- lapply(y, as.character)
y[] <- lapply(y, as.numeric)
y$arm <- as.factor(y$arm)
y$site <- as.factor(y$site)
y$Subject <- as.factor(y$Subject)
y \%>\% metaplot(conc)
#y \%>\% metaplot(site)
y \%>\% metaplot(Wt, arm)
y \%>\% metaplot(arm, Wt)
y \%>\% metaplot(arm, Wt,site)
y \%>\% metaplot(conc, Time)
#y \%>\% metaplot(arm, site)
y \%>\% metaplot(conc, Time, Subject)
y \%>\% metaplot(conc, Time, , Subject)
y \%>\% metaplot(conc, Time, Subject, site)
y \%>\% metaplot(conc, Time, Subject, site, arm)
y \%>\% metaplot(lKe, lKa, lCl)
y \%>\% scatter(conc, ipred, Time)
y \%>\% scatter(conc, ipred, Time, Subject)
x \%>\% metaplot(conc, ipred, Time, Subject, colors = 'black', points = c(T,F), lines = c(F,T))
y \%>\% scatter(conc, ipred, Time, site, arm)
y \%<>\% mutate(Time = ifelse(Time > 15, NA, Time))
y \%>\% scatter(conc, ipred, Time, site, arm)
}}

}
\seealso{
Other metaplot: \code{\link{boxplot.data.frame}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{densplot}}, \code{\link{metaplot}},
  \code{\link{scatter}}

Other univariate plots: \code{\link{densplot.data.frame}},
  \code{\link{densplot_data_frame}}, \code{\link{densplot}}

Other bivariate plots: \code{\link{scatter.data.frame}},
  \code{\link{scatter_data_frame}}, \code{\link{scatter}}

Other multivariate plots: \code{\link{corsplom.data.frame}},
  \code{\link{corsplom_data_frame}}
}
