% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot.folded}
\alias{metaplot.folded}
\title{Metaplot Method for Folded}
\usage{
\method{metaplot}{folded}(x, ...)
}
\arguments{
\item{x}{folded}

\item{...}{unquoted anonymous arguments are passed as character: var to metaplot_().}
}
\description{
Plots a folded object.
}
\seealso{
\code{\link{metaplot_.folded}}
}
