% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot}
\alias{metaplot}
\title{Metaplot}
\usage{
metaplot(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a metaplot. Generic.  See \code{\link{metaplot.data.frame}}.
}
\examples{
example(metaplot.data.frame)
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{densplot}},
  \code{\link{scatter}}

Other metaplot: \code{\link{boxplot.data.frame}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{densplot}},
  \code{\link{metaplot.data.frame}}, \code{\link{scatter}}
}
