% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot}
\alias{metaplot}
\title{Metaplot}
\usage{
metaplot(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Creates a metaplot.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{continuous}}, \code{\link{corsplom}},
  \code{\link{dens}}, \code{\link{guide}},
  \code{\link{label}}, \code{\link{overlay}},
  \code{\link{scatter}}

Other metaplot: \code{\link{metaplot.data.frame}},
  \code{\link{metaplot.folded}},
  \code{\link{metaplot.grouped_df}}
}
