% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPlot.R
\name{metaPlot}
\alias{metaPlot}
\title{MetaPlot Function}
\usage{
metaPlot(
  bed_file = NULL,
  group_names = NULL,
  features_anno = NULL,
  scale_region = FALSE,
  cut_ratio = NULL,
  facet = FALSE,
  geom_density_bw = NULL,
  geom_density_n = 512,
  col_alpha = 0.5,
  line_color = "#CC0033",
  line_width = 1
)
}
\arguments{
\item{bed_file}{A list of BED file paths or data frames containing regions of interest. Default is \code{NULL}.}

\item{group_names}{A vector of names corresponding to each BED file for grouping in the plot. Default is \code{NULL}, automatically set to \code{bed_file}.}

\item{features_anno}{A features_anno object or file path used to map features. Default is \code{NULL}.}

\item{scale_region}{Logical, whether to scale regions relative to feature lengths. Default is \code{FALSE}.}

\item{cut_ratio}{A numeric value indicating the ratio to cut the regions for scaling. Default is \code{NULL}.}

\item{facet}{Logical, whether to use faceting in the plot for separate groups. Default is \code{FALSE}.}

\item{geom_density_bw}{Numeric, specifying the bandwidth for \code{geom_density}. Default is \code{NULL}, uses automatic bandwidth selection.}

\item{geom_density_n}{Integer, number of equally spaced points at which the density is estimated. Default is \code{512}.}

\item{col_alpha}{Numeric, alpha value for color transparency in the density plot. Default is \code{0.5}.}

\item{line_color}{A color code for lines in the density plot if there is only one bed file. Default is \code{"#CC0033"}.}

\item{line_width}{Line width for density curve. Default is \code{1}.}
}
\value{
A list with two components:
\describe{
\item{data}{A combined data frame with relative positions of peaks against annotated features.}
\item{plot}{A ggplot2 object representing the density plot of peaks across the specified genomic features.}
}
}
\description{
Generate a meta plot of peak densities across genomic features such as UTRs and CDS using BED files.
}
