% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaPlot.R
\name{metaPlot}
\alias{metaPlot}
\title{metaplot}
\usage{
metaPlot(
  bed_file = NULL,
  group_names = NULL,
  features_anno = NULL,
  scale_region = FALSE,
  cut_ratio = NULL,
  facet = FALSE
)
}
\arguments{
\item{bed_file}{bed format of peaks files.}

\item{group_names}{vector of group names.}

\item{features_anno}{features_anno object from prepare_features functions.}

\item{scale_region}{whether do scale for each feature length, default FALSE.}

\item{cut_ratio}{whether use custom ratio to plot feature regions,
for eaxample: c(0.1,0.8,0.1).}

\item{facet}{whether dispaly plot for each group, default FALSE.}
}
\value{
A list with data and plot
}
\description{
metaplot
}
