% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{metaplot_}
\alias{metaplot_}
\title{Metaplot, Standard Evaluation}
\usage{
metaplot_(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}

\item{var}{character: quoted names of variables to plot}
}
\description{
Creates a metaplot using standard evaluation.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{continuous}}, \code{\link{corsplom}},
  \code{\link{dens}}, \code{\link{guide}},
  \code{\link{label}}, \code{\link{metaplot}},
  \code{\link{scatter}}
}
\keyword{internal}
