% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metaplot_ref}
\alias{metaplot_ref}
\title{Calculate Reference Values}
\usage{
metaplot_ref(x, var, ...)
}
\arguments{
\item{x}{data.frame}

\item{var}{name of vector in x}

\item{...}{ignored}
}
\value{
numeric
}
\description{
Calculates reference values for x and y axes.  Coerces column attribute 'reference' to numeric: a single value or an encoding giving multiple numeric values (decodes are ignored).
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_gg_scatter}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_key}},
  \code{\link{panel.meta_densityplot}},
  \code{\link{panel_tile}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other reference lines: \code{\link{diag_pin}},
  \code{\link{scatter_panel_ref}}
}
