% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metastats}
\alias{metastats}
\title{Format GLM Statistics}
\usage{
metastats(
  x,
  y,
  family = if (all(y \%in\% 0:1, na.rm = TRUE)) "binomial" else "gaussian",
  ...
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{family}{regression family}

\item{...}{other arguments}
}
\value{
character
}
\description{
Formats GLM statistics. Uses a gaussian family by default, or binomial family if all y are 0 or 1, to fit a general linear model.  Formats number of observations, p-value, and Pearson correlation coefficient into a string for printing.
}
\seealso{
\code{\link{scatter_panel}}

Other regression functions: 
\code{\link{model}()},
\code{\link{region}()}
}
\concept{regression functions}
