% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{my.diag.panel}
\alias{my.diag.panel}
\title{Diagonal Panel Function}
\usage{
my.diag.panel(x, denscale = 0.2, ...)
}
\arguments{
\item{x}{numeric}

\item{denscale}{inflation factor for height of density smooth}

\item{...}{passed arguments}
}
\description{
Diagonal panel function for corsplom(). Plots a density smooth against the corresponding axis from within the diagonal panel.  Plots a grey pin at each axis zero.
}
\seealso{
Other panel functions: \code{\link{l.p}}, \code{\link{u.p}}
}
\keyword{internal}
