% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{pack}
\alias{pack}
\title{Pack Something}
\usage{
pack(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Pack Something.  Generic, with method for data.frame.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{densplot}}, \code{\link{metaplot}},
  \code{\link{scatter}}, \code{\link{test_metaplot}},
  \code{\link{unpack}}

Other pack: \code{\link{pack.data.frame}},
  \code{\link{unpack.data.frame}}, \code{\link{unpack}}
}
