% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical.R
\name{panel_tile}
\alias{panel_tile}
\title{Draw a Tile}
\usage{
panel_tile(x, y, subscripts, group.number, group.value, fill.alpha, line.alpha,
  col, col.line, loc, msg, .src, alpha, cex, ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{subscripts}{subscripts}

\item{group.number}{group number}

\item{group.value}{group value}

\item{fill.alpha}{alpha transparency for fill}

\item{line.alpha}{alpha transparency for line}

\item{col}{fill color}

\item{col.line}{border color}

\item{loc}{location for output of msg}

\item{msg}{ignored}

\item{.src}{data source for which subscripts give x, y, msg,  and tile limits}

\item{alpha}{ignored}

\item{cex}{expansion for msg text; passed to msg}

\item{...}{passed arguments}

\item{tex}{tile expansion: scale factor for reducing each tile size relative to full size (<= 1)}
}
\description{
Draws a tile in a mosaic.
}
\seealso{
Other panel functions: \code{\link{boxplot_panel}},
  \code{\link{categorical_panel}},
  \code{\link{corsplom_gg_correlation}},
  \code{\link{corsplom_gg_diagonal}},
  \code{\link{corsplom_gg_scatter}},
  \code{\link{corsplom_panel_correlation}},
  \code{\link{corsplom_panel_diagonal}},
  \code{\link{corsplom_panel_scatter}},
  \code{\link{dens_panel}}, \code{\link{diag_label}},
  \code{\link{diag_pin}}, \code{\link{iso_prepanel}},
  \code{\link{metaplot_ref}},
  \code{\link{scatter_panel_ref}},
  \code{\link{scatter_panel}}

Other categorical: \code{\link{categorical.data.frame}},
  \code{\link{categorical_data_frame}},
  \code{\link{categorical_panel}},
  \code{\link{categorical}}
}
