% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCplot.R
\name{ROCplot}
\alias{ROCplot}
\title{Draws ROC curves for diagnostic accuracy studies}
\usage{
ROCplot(source_tbl, study, TP = "TP", FN = "FN", FP = "FP",
  TN = "TN", group_var = NULL, marker = NULL, colours = NULL,
  fontsize = 8, plabs = TRUE, axlabs = NULL)
}
\arguments{
\item{source_tbl}{Dataframe containing the meta-analysis data}

\item{study}{Index column for unique study identifier e.g. "Obelix 50 B.c."}

\item{TP}{Column  name for column containing the True Positive values}

\item{FN}{Column  name for column containing the False Negative values.}

\item{FP}{Column  name for column containing the False Positive values.}

\item{TN}{Column  name for column containing the True Negative values.}

\item{group_var}{Name of the column containig group identifiers for setting marker shape. Leave empty if not required.}

\item{marker}{Name of the column containing values for setting marker size. Leave empty if not required.}

\item{colours}{List of colours for group_var. Standard is greyscale.}

\item{fontsize}{Font Size}

\item{plabs}{Should each point be labled?}

\item{axlabs}{X and Y axis lables, given as a list: c("X","Y")}
}
\value{
none
}
\description{
This function creates an ROC plot. Returns or prints a ggplot object.
}
