% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter.data.frame}
\alias{scatter.data.frame}
\title{Scatterplot Method for Data Frame}
\usage{
\method{scatter}{data.frame}(x, ..., fun = getOption("metaplot_scatter",
  "scatter_data_frame"))
}
\arguments{
\item{x}{data.frame}

\item{...}{passed to fun}

\item{fun}{function to draw the plot}
}
\description{
Scatterplot method for class 'data.frame'. Parses arguments and generates the call: fun(x, yvar, xvar, groups, facets, ...).
}
\examples{
library(magrittr)
library(dplyr)
attr(Theoph$conc,'label') <- 'theophylline concentration'
attr(Theoph$conc,'guide') <- 'mg/L'
attr(Theoph$Time,'label') <- 'time'
attr(Theoph$Time,'guide') <- 'h'
attr(Theoph$Subject,'guide') <- '////'
scatter(Theoph,conc, Time)
scatter(Theoph, conc, Time, Subject) # Subject as groups
scatter(Theoph, conc, Time, , Subject) # Subject as facet
scatter(Theoph \%>\% filter(conc > 0), conc, Time, Subject, ylog = TRUE, yref = 5)
scatter(Theoph, conc, Time, Subject,ysmooth = TRUE)
scatter(Theoph, conc, Time, main = TRUE, corr = TRUE)
scatter(Theoph, conc, Time, conf = TRUE, loc = 3, yref = 6)
}
\seealso{
\code{\link{scatter_data_frame}}

Other bivariate plots: \code{\link{metaplot.data.frame}},
  \code{\link{scatter_data_frame}}

Other scatter: \code{\link{scatter_data_frame}}
}
