% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaplot.R
\name{scatter.data.frame}
\alias{scatter.data.frame}
\title{Scatterplot for Data Frame}
\usage{
\method{scatter}{data.frame}(x, .y, .x, groups = NULL, ..., ylog = FALSE,
  xlog = FALSE, yref = NULL, xref = NULL, ysmooth = FALSE,
  xsmooth = FALSE, cols = 3, density = FALSE, iso = FALSE,
  main = TRUE, group_codes = NULL, crit = 1.3)
}
\arguments{
\item{x}{data.frame}

\item{.y}{y axis item}

\item{.x}{x axis item}

\item{groups}{optional grouping item}

\item{...}{passed arguments}

\item{ylog}{log transform y axis (guessed if missing)}

\item{xlog}{log transform x axis (guessed if missing)}

\item{yref}{reference line from y axis}

\item{xref}{reference line from x axis}

\item{ysmooth}{supply loess smooth of y on x}

\item{xsmooth}{supply loess smmoth of x on y}

\item{cols}{suggested columns for auto.key}

\item{density}{plot point density instead of points}

\item{iso}{use isometric axes with line of unity}

\item{main}{print Pearson correlation coefficient as title}

\item{group_codes}{append these to group values for display purposes}

\item{crit}{if ylog or xlog missing, log transform if mean/median ratio for non-missing values is greater than crit}
}
\description{
Scatterplot for class 'data.frame'.
}
