% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatterplot}
\usage{
scatter(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Scatterplot.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{densplot}},
  \code{\link{metaplot}}

Other scatter: \code{\link{metapanel}},
  \code{\link{scatter.data.frame}},
  \code{\link{scatter_data_frame}}

Other bivariate plots: \code{\link{metaplot.data.frame}},
  \code{\link{scatter.data.frame}},
  \code{\link{scatter_data_frame}}

Other metaplot: \code{\link{boxplot.data.frame}},
  \code{\link{categorical}}, \code{\link{corsplom}},
  \code{\link{densplot}},
  \code{\link{metaplot.data.frame}}, \code{\link{metaplot}}
}
