% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatterplot}
\usage{
scatter(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Scatterplot.
}
\seealso{
Other generic functions: \code{\link{axislabel}},
  \code{\link{corsplom}}, \code{\link{dens}},
  \code{\link{guide}}, \code{\link{label}},
  \code{\link{metaplot}}
}
