% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatterplot}
\usage{
scatter(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Scatterplot.
}
\seealso{
Other generic functions: 
\code{\link{axislabel}()},
\code{\link{categorical}()},
\code{\link{corsplom}()},
\code{\link{densplot}()},
\code{\link{metaplot}()},
\code{\link{pack}()},
\code{\link{test_metaplot}()},
\code{\link{unpack}()}

Other scatter: 
\code{\link{metaplot_key}()},
\code{\link{scatter.data.frame}()},
\code{\link{scatter_data_frame}()},
\code{\link{scatter_panel}()}

Other bivariate plots: 
\code{\link{iso_prepanel}()},
\code{\link{metaplot.data.frame}()},
\code{\link{scatter.data.frame}()},
\code{\link{scatter_data_frame}()}
}
\concept{bivariate plots}
\concept{generic functions}
\concept{scatter}
