% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{u.p}
\alias{u.p}
\title{Upper Panel Function}
\usage{
u.p(x, y, col = "black", loess = "red", ...)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{col}{point color}

\item{loess}{loess color}

\item{...}{passed arguments}
}
\description{
Upper panel function for corsplom(). Plots data with loess smooth.
}
\seealso{
Other panel functions: \code{\link{l.p}},
  \code{\link{metapanel}}, \code{\link{my.diag.panel}}
}
\keyword{internal}
