% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wikisym2plotmath}
\alias{wikisym2plotmath}
\title{Convert Wiki Symbol to Plotmath}
\usage{
wikisym2plotmath(x, ...)
}
\arguments{
\item{x}{character}

\item{...}{ignored}
}
\value{
expression
}
\description{
Converts wiki symbol to plotmath.  Vectorized version of \code{\link{wikisym2plotmathOne}}.
}
\seealso{
Other formatters: \code{\link{diag_label}},
  \code{\link{wikisym2plotmathOne}}
}
